% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemePack.R
\name{addFootnote}
\alias{addFootnote}
\title{Function footnote}
\usage{
addFootnote(
  my.ggp,
  footnoteText,
  footnoteSize,
  footnoteColor,
  footnoteJust = 1,
  yoffset = 0
)
}
\description{
Function footnote
}
