% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemePack.R
\name{baseFont}
\alias{baseFont}
\title{Function  baseFont}
\usage{
baseFont(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Size for the basefont in points (Default = 12)}

\item{base_family}{Set the font family}
}
\value{
A ggplot theme that can be added to a plot object
}
\description{
Sets just the font size and optionally font family without changing other
graphic attributes.  Notably, the ggplot theme_grey and theme-bw can be used
to change base_size but annoyingly reset my custom legend locations. So apply
baseFont last and use it to change just the font sizes without otherwise
altering you plot theme.
}
\details{
Optionally, set the font family as well. But from what I can tell,
on Windows, you can have any font you want as long as it's Helvetica.
see: \url{https://github.com/wch/extrafont}
regarding attempting to use other font families.
}
\examples{
Myggplot <- myggplot + baseFont(18)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{fontsize}
\keyword{ggplot2}
