% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemePack.R
\name{baseTheme}
\alias{baseTheme}
\title{Function  baseTheme}
\usage{
baseTheme(base_size = 18, base_family = "", scale_legend = TRUE)
}
\arguments{
\item{base_size}{Size for the basefont in points (Default = 18)}

\item{base_family}{Set the font family}

\item{scale_legend}{Scales Legend.Text to 10/base_size lines (Default=TRUE)}
}
\value{
A ggplot theme that can be added to a plot object
}
\description{
A basic theme for individual plots that only sets relative font sizes for
common graphic elements.  Apply baseTheme as your last layer.  In particular,
theme_bw, theme_grey, will undo some of the changes set by baseTheme.  After
applying baseTheme, you can use baseFont to adjust font sizes for different
output purposes (e.g. larger for PPT, smaller for knitr).
}
\examples{
Myggplot = myggplot + baseTheme(12)  #start with a 12pt font theme
Myggplot = myggplot + baseFont(24)  #change plot to 24pt base.

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{fontsize}
\keyword{ggplot2}
