% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_RSE.R
\name{Build_RSE}
\alias{Build_RSE}
\title{Function  Build_RSE}
\usage{
Build_RSE(.GeneData, OutputPath = ".")
}
\arguments{
\item{OutputPath}{Path}

\item{GeneData}{A predefined List of Lists defining data names, file names, and file types.}

\item{PlotFile}{Name for image of sample zFPKM distribution fits (set to NULL to disable plot)}

\item{debug}{If TRUE, returns a simple list of data objects instead of building
a RSE object.  Use this to examine the data if the SummarizedExperiment
function issues errors.  You can then examine the dataframes that generated
the errors. Use debug mode if you get errors after the message "Building RSE Object".}
}
\value{
A RangedSummarizedExperiment
}
\description{
Build a RangedSummarizedExperiment (RSE) from count data.  The
RSE is returned.
}
\details{
User provides separate tab-delimited text files for each row annotation,
column anotation and "assays" (matrices of RxC e.g. counts, FPKM, etc).

GeneData is a predefined data structure (list of lists) that specifies the
datafiles to import.  GeneData is pre-configured for the standarized Omicsoft
project output.  Modify as necessary to accommodate data from other
pipelines.  There is also a "TranscriptData" pre-defined definition.  Passing
TranscriptData as the GeneData parameter will produce a
Transcript-level RSE.  Note a single RSE cannot contain both Gene and Transcript
data because of the mismatch in gene/transcript rowcounts.

Data Requirements:

Minimally, Sample (column) annotation, gene or transcript (row) annotation and
counts are required to build an RSE object.

Column 1 in annotation and assay data files should be named GeneID or
TranscriptID and the ID order in all files should be the same.

Similarly, the row order of the Column annotation should match the
column order in the assay data files.

The Gene/Transcript annotation must include chromosome position data (chr,
start, end, strand).  FPKM and TPM will be imported if available or
calculated and added to the RSE if an
ExonLength field is provided in the row annotation.  Alignment QC data
will also be imported, if present, and placed in the metadata slot.
}
\examples{
MyRSE = Build_RSE (GeneData, PlotFile = "Gene.zFPKM.PNG")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Entrez,}
\keyword{GeneID,}
\keyword{Omicsoft}
\keyword{SummarizedExperiment,}
\keyword{gene}
\keyword{symbol,}

