% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildAndSaveRSE.R
\name{BuildAndSaveRSE}
\alias{BuildAndSaveRSE}
\title{Function  BuildAndSaveRSE}
\usage{
BuildAndSaveRSE(InputFolder = ".", OutputFolder = "./RData",
  RSEfilename = "RSE.RDS", Species = "", PlotFilename = "zFPKM.PNG",
  Level = "GeneData", FacetTitles = FALSE)
}
\arguments{
\item{InputFolder}{defaults to current folder}

\item{OutputFolder}{defaults to "./RData"}

\item{RSEfilename}{Filename for the RSE.RDS output file}

\item{Species}{One of Human, Mouse or Rat.}

\item{PlotFilename}{Filename for the zFPKM PNG file.}

\item{Level}{One of "GeneData" or "TranscriptData" (defaults to "GeneData")}

\item{FacetTitles}{Name for image of sample distribution fits}
}
\value{
Datafram with Entez GeneID column added
}
\description{
Build a RangedSummarizedExperiment (RSE) from Omicsoft Alignment and count data.  The
RSE is saved as a .RDS file.  Users provides
separate tab-delimited text files for each row annotation, column anotation
and "assay" (matrices of RxC e.g. counts, FPKM). GeneData is a predefined data
structure (list of lists) that specifies the datafiles to import.
}
\examples{
MyDataframe = Build_RSE (GeneData, Species = "Human", OutputPath = "./RData", PlotFile = "Gene.zFPKM.PNG")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Entrez,}
\keyword{GeneID,}
\keyword{Omicsoft}
\keyword{SummarizedExperiment,}
\keyword{gene}
\keyword{symbol,}

