% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemePack.R
\name{bwTheme}
\alias{bwTheme}
\title{Function  bwTheme}
\usage{
bwTheme(base_size = 12, base_family = "", scale_legend = TRUE)
}
\arguments{
\item{base_size}{Size for the basefont in points (Default = 12)}

\item{base_family}{Set the font family}

\item{scale_legend}{Scales Legend.Text to 10/base_size lines (Default=TRUE)}
}
\value{
A ggplot theme that can be added to a plot object
}
\description{
A customization of the theme_grey default ggplot theme.  This theme establishes
relative sizes for most elements (title, axis labels and legends) so that it is
easy to adjust all font sizes without otherwise altering you plot by using the
baseFont function.  I also use a darker color for the axis scale labels (grey33 instead of grey50).
I like the idea of having the axis title slightly darker than the axis numbering but
thought HW's choice of grey50 for the axis numbering was too light.
}
\examples{
Myggplot = myggplot + bwTheme(18)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{fontsize}
\keyword{ggplot2}
