% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSex.R
\name{checkSex}
\alias{checkSex}
\title{Function checkSex}
\usage{
checkSex(dgeObj, species, sexCol, labelCol, showLabels = FALSE, chrX = "X",
  chrY = "Y", baseFontSize = 14)
}
\arguments{
\item{dgeObj}{A DGEobj (Required)}

\item{species}{One of "human", "mouse", "rat"}

\item{sexCol}{Name of the sex annotation column in the design table within the dgeObj}

\item{labelCol}{Name of the design column to use to label points}

\item{chrX}{Name of the X chromosome in the gene annotation within the geneData object (Default = "X")}

\item{chrY}{Name of the Y chromosome in the gene annotation within the geneData object (Default = "Y")}

\item{baseFontSize}{Sets the base font size for ggplot themes (Default = 14)}

\item{showLables}{Set TRUE to label points with labelcol (default = FALSE)}
}
\value{
A ggplot object
}
\description{
Take a DGEobj as input and plot expression of XIST vs the highest expressed Y chr gene.
}
\examples{

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGE,}
\keyword{MDS,}
\keyword{QC}
\keyword{RNA-Seq,}
