% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCounts.R
\name{convertCounts}
\alias{convertCounts}
\title{Function convertCounts}
\usage{
convertCounts(counts, unit, geneLength, log = FALSE, normalize = FALSE,
  PlotDir, PlotFile, FacetTitles)
}
\arguments{
\item{counts}{A numeric matrix or dataframe of N genes x M Samples.  All columns
must be numeric.}

\item{unit}{Required. One of CPM, FPKM, zFPKM, FPK or TPM.}

\item{geneLength}{Required for length-normalizes units (TPM, FPKM, zFPKM or FPK)}

\item{log}{Default = FALSE.  Set TRUE to return Log2 values. Log conversion
employs the edgeR method which uses an average prior of 0.25 moderated by the
   library size.}

\item{normalize}{Default = FALSE. TRUE activates TMM normalization.
Other allowed values are: "TMM","RLE","upperquartile".  Invokes
edgeR::calcNormFactors for normalization.}

\item{PlotDir}{Only applies to zFPKM. Specifies folder for zFPKM distribution plot.}

\item{PlotFile}{Only applies to zFPKM. Default = "zFPKM_Distribution.png"}

\item{FacetTitles}{Only applies to zFPKM. Turn facet titles on or off
(Default = TRUE)}
}
\value{
A matrix in the new unit space
}
\description{
Takes a count matrix as input and converts to other desired units.  Supported
units include CPM, FPKM, FPK, zFPKM and TPM.  Output units can be logged
and/or normalized.  Calculations are performed using edgeR functions except
for the conversion to TPM which is converted from FPKM using the formula provided
by [Harold Pimental](https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/).
}
\examples{
#TMM normalized Log2TPM
Log2TPM = convertCounts(mycounts),
                      unit="TPM",
                      geneLength=gene.annotation$ExonLength,
                      log=TRUE,
                      normalize=TRUE)

#un-normalized CPM (not logged)
RawCPM = convertCounts(MyCounts,
                      unit="CPM",
                      log=FALSE,
                      normalize=FALSE)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq,}
\keyword{conversion}
\keyword{unit}
