% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGE_Unit_Conversion.R
\name{counts2TMM_TPM}
\alias{counts2TMM_TPM}
\title{Function counts2TMM_TPM:  Convert Counts to TMM normalized TPM}
\usage{
counts2TMM_TPM(counts, genelength)
}
\arguments{
\item{counts}{a matrix of genes(rows) by samples(col); use all genes, not a filtered gene list}

\item{genelength}{a required vector of gene lengths in the same order as rows of counts.  This is typically the effective length
 from the RSEM calculation.  You can get this from the "exonlength" column in Omicsoft Annotation and the effective
length data should also be available for Xpress projects as of Aug2016.}
}
\value{
matrix of TMM normalized TPM values
}
\description{
TMM alone does not capture the compositional bias normalization provided by TMM normalization.
This function uses edgeR functions to calculate TMM normalized CPM and then coverts it to TMM units.
Note that you should use a complete count matrix of all genes for the TPM units to be correct as
the conversion to TPM involves a column sum.
}
\details{
Reference = https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/
Modified by JRT to work with an input matrix instead of just a vector.
}
\examples{
MyTPM = counts2TMM_TPM(MyCounts, MyGeneLengths)

}
\author{
https://haroldpimentel.wordpress.com/
}
\keyword{FPKM}
\keyword{TPM;}
\keyword{counts;}
\keyword{read}
