% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGE_Unit_Conversion.R
\name{countToFpkm}
\alias{countToFpkm}
\title{Function countToFpkm:  Convert Counts matrix to TPM}
\usage{
countToFpkm(counts, effLen)
}
\arguments{
\item{counts}{a matrix of genes(rows) and samples(col)}

\item{effLen}{vector or gene lengths (length must equal length(counts)).}
}
\value{
matrix of FPKM values
}
\description{
To get a proper FPKM value you should use the unfiltered FPKM matrix as input to this function.
This is because a column sum is involved in the TPM calculation and the column sum should include all mapped reads.
}
\details{
Reference = https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/
Modified by JRT to work with an input matrix instead of just a vector.
}
\examples{
MyFPKM = countToTpm(counts, MyCounts)

}
\author{
https://haroldpimentel.wordpress.com/
}
\keyword{TPM}
\keyword{counts;}
\keyword{read}
