% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGE_Unit_Conversion.R
\name{cpmToFpkm}
\alias{cpmToFpkm}
\title{Function cpmToFpkm:  Convert Counts to TPM}
\usage{
cpmToFpkm(CPM, exonLength)
}
\arguments{
\item{counts}{a matrix of genes(rows) and samples(col)}

\item{effLen}{vector or gene lengths (length must equal length(counts)).}
}
\value{
matrix of FPKM values
}
\description{
Reference = https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/
}
\examples{
MyFPKM = cpmToFpkm(MyCPM, exonlengths)

}
\author{
https://haroldpimentel.wordpress.com/
}
\keyword{CPM;}
\keyword{FPKM}

