% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_excel.R
\name{df_to_excel}
\alias{df_to_excel}
\title{Function  df_to_excel}
\usage{
df_to_excel(
  df,
  filename,
  sheetname = "data",
  widthFirstCol = 40,
  widthOtherCol = 10,
  baseFontSize = 10,
  useRownames = TRUE
)
}
\arguments{
\item{df}{A dataframe with rownames and colnames (Required).}

\item{filename}{Path/Filename for output (Required).}

\item{sheetname}{Name for the Excel sheet (Default = "data")}

\item{widthFirstCol}{width of 1st column (rownames) (Default=40)}

\item{widthOtherCol}{width of 1st column (rownames) (Default=10)}

\item{baseFontSize}{base font for Excel file (Default=10)}

\item{useRownames}{Creates a spreadsheet without rownames (default=TRUE).  Set to FALSE
when there is no column suitably unique to use for rownames}
}
\value{
Writes an Excel file to the specified file.
}
\description{
Take a dataframe with rownames and output to an Excel spreadsheet.
Freezes the row and colnames and sets columns into Excel filter mode.
}
\examples{
#merge a set of topTable contrasts into one data.frame
MyContrastTable  = topTable.merge (topTablelist)
df_to_excel(MyContrastTable, "myExcelFile.xlsx")

}
\author{
John Thompson, \email{jrt@thompsonclan.org}
}
\keyword{Excel}
\keyword{dataframe,}
