% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_excel.R
\name{dflist_to_excel}
\alias{dflist_to_excel}
\title{Function  dflist_to_excel}
\usage{
dflist_to_excel(
  dflist,
  filename,
  widthFirstCol = 40,
  widthOtherCol = 10,
  baseFontSize = 10,
  useRownames = TRUE
)
}
\arguments{
\item{dflist}{A named list of dataframes with rownames and colnames (Required).}

\item{filename}{Path/Filename for output (Required).}

\item{widthFirstCol}{width of 1st column (rownames) (Default=40)}

\item{widthOtherCol}{width of 1st column (rownames) (Default=10)}

\item{baseFontSize}{base font for Excel file (Default=10)}

\item{useRownames}{default = TRUE; Set to false when no appropriate unique column present.}
}
\value{
Writes an Excel file to the specified file.
}
\description{
Take a named list of dataframes with rownames and output to an Excel spreadsheet.
Freezes the row and colnames and sets columns into Excel filter mode.
}
\examples{
MyContrastTable  = topTable.merge (topTablelist)

}
\author{
John Thompson, \email{jrt@thompsonclan.org}
}
\keyword{Excel}
\keyword{dataframe,}
