% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEtools.R
\docType{package}
\name{DGEtools}
\alias{DGEtools}
\title{DGEtools}
\description{
Provides a function toolkit to facilitate reproducible RNA-Seq DGE
analysis by standardizing several types of common plots and defining functions that
facilitate common differential gene expression analysis tasks.
}
\section{Function Overview}{


TBD
}

\section{More Information}{

\code{browseVignettes(package = 'DGEtools')}
}

