% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCol.R
\name{.extractCol2}
\alias{.extractCol2}
\title{Function  .extractCol2}
\usage{
.extractCol2(dflist, colName)
}
\arguments{
\item{dflist}{A named list of dataframes which all have some common colnames and
some overlap on rownames (genes). and same row counts. The dataframes in
the list should have rownames (geneIDs).}

\item{colName}{The name of the data column to extract to a matrix.  This
column must be present in all dataframes on the list.}
}
\value{
A dataframe containing the extracted columns
}
\description{
Take a named list of dataframes where each dataframe has the same
column names (e.g. a list of topTable dataframes). Extract
the named column from each dataframe and return a matrix.
}
\details{
The common use case for this is to provide a list of topTable
data frames and extract one column from each file to create
a matrix of LogRatios or Pvalues.
}
\examples{

    MyPvalues  = .extractCol2(TopTableList, "P.Value")

}
\author{
John Thompson, \email{jrt@thompsonclan.org}
}
\keyword{topTable}
