% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblGene2Entrez.R
\name{EnsemblGene2Entrez}
\alias{EnsemblGene2Entrez}
\title{Function  EnsemblGene2Entrez}
\usage{
EnsemblGene2Entrez(DF, Species = NULL, EnsemblGene = "GeneID")
}
\arguments{
\item{DF}{a dataframe with a Ensembl gene ID column}

\item{Species}{One of Human, Mouse, Rat or Dog.}

\item{EnsemblGene}{Column name of the Ensembl GeneID column to map}
}
\value{
Dataframe with Entez GeneID column added
}
\description{
Map Ensembl GeneIDs to Entrez GeneIDs using org.Xx.eg.db packages
for Human, Mouse, Rat and Dog
}
\details{
One to Many handling: If the EnsemblID maps to multiple EntrezIDs, 
the EntrezID field will contain a comma-separated
list of EntrezIDs.
}
\examples{
MyDataframe = EnsemblGene2Entrez (MyDataframe, Species = "Human", EnsemblGene = "GeneID")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Entrez,}
\keyword{GeneID}
\keyword{gene}
\keyword{symbol,}
