% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblGene2GeneSym.R
\name{EnsemblGene2GeneSym}
\alias{EnsemblGene2GeneSym}
\title{Function  EnsemblGene2GeneSym}
\usage{
EnsemblGene2GeneSym(DF, Build = NULL, MergeAll = FALSE)
}
\arguments{
\item{DF}{a dataframe with a gene symbol column}

\item{Build}{One of grch38, grch37, grcm38, rnor6.}

\item{MergeAll}{If TRUE, will merge all annotation columns from annotables.
If FALSE (default), will only merge symbol, biotype and description columns.}
}
\value{
Dataframe with Entez GeneID, biotype and desscripton columns added
}
\description{
Takes a dataframe from topTable or topTreat and maps Ensembl GeneIDs to Gene
Symbols using the Annotables package.  Technically it should work on
any dataframe with Ensembl Gene IDs as rowlabels.  The added columns to the
include "symbol", "biotype" and "description".
}
\examples{
#Annotate one topTable dataframe
MyDataframe = EnsemblGene2GeneSym (MyDataframe, Build = "grch38")

#same with less typing
MyDataframe \%<>\% EnsemblGene2GeneSym("grch38")

#Annotate a whole bunch of topTable dataframes in a contrast list
#produced by runContrasts
for (i in 1:length(ContrastList$TopTableList)){
   ContrastList$TopTableList[[i]] \%<>\% EnsemblGene2GeneSym("grcm38")
}

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Ensembl,}
\keyword{GeneID}
\keyword{gene}
\keyword{symbol,}

