% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entrez2GeneSym.R
\name{Entrez2GeneSym}
\alias{Entrez2GeneSym}
\title{Function Entrez2GeneSym:  Map Official Gene Symbols to Entrez GeneIDs}
\usage{
Entrez2GeneSym(DF, Species = NULL, EntrezCol = "EntrezID")
}
\arguments{
\item{DF}{a dataframe with a Entrez ID column}

\item{Species}{One of "Human", "Mouse", "Rat" or "Dog" (required)}

\item{EntrezCol}{Column name of the Entrez gene ID column to map}
}
\value{
DataFrame with Entrez GeneID column added
}
\description{
Maps Entrez GeneID to Gene Symbols using the org.Xx.eg.db packages
for Human, Mouse and Rat. Aggregates duplicate mappings (genesym mapped to 2
or more Entrez IDs) to avoid row expansion.
}
\details{
Note: currently aggregate mappings are lost due to exact matching.  Need
to implement a second pass for unmapped genesymbols to use partial matching.
}
\examples{
MyDataframe = GeneSym2Entrez (MyDataframe, Species = "Human", EntrezCol = "GeneSym")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Entrez,}
\keyword{GeneID}
\keyword{gene}
\keyword{symbol,}

