% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntrezGene2Ensembl.R
\name{EntrezGene2Ensembl}
\alias{EntrezGene2Ensembl}
\title{Function  EntrezGene2Ensembl}
\usage{
EntrezGene2Ensembl(DF, Species = NULL, EntrezGene = "GeneID")
}
\arguments{
\item{DF}{a dataframe with a Ensembl gene ID column}

\item{Species}{One of Human, Mouse, Rat or Dog.}

\item{EntrezGene}{Column name of the Entrez GeneID column to map}
}
\value{
Dataframe with Entez GeneID column added
}
\description{
Map Entrez GeneIDs to Ensembl GeneIDs using org.Xx.eg.db packages
for Human, Mouse, Rat or dog
}
\details{
One to Many handling: If the EntrezID maps to multiple EnsemblIDs, 
the EnsemblID field will contain a comma-separated
list of matching EnsemblIDs.
}
\examples{
MyDataframe = EnsemblGene2Entrez (MyDataframe, Species = "Human", EnsemblGene = "GeneID")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Entrez,}
\keyword{GeneID}
\keyword{gene}
\keyword{symbol,}

