% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemePack.R
\name{facetTheme}
\alias{facetTheme}
\title{Function  facetTheme}
\usage{
facetTheme(base_size = 18, base_family = "")
}
\arguments{
\item{base_size}{Size for the basefont in points (Default = 18)}

\item{base_family}{Set the font family}
}
\value{
A ggplot theme that can be added to a plot object
}
\description{
A basic theme for facetted plots that sets relative font sizes
for common graphic elements.  Use in conjunction with baseFont
to easily adjust all fonts in a plot.
}
\examples{
Myggplot = myggplot + facetTheme(12)  #start with a 12pt font theme
Myggplot = myggplot + baseFont(24)  #change plot to 24pt base.

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{fontsize}
\keyword{ggplot2}

