% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSym2Entrez.R
\name{GeneSym2Entrez}
\alias{GeneSym2Entrez}
\title{Function GeneSym2Entrez:  Map Official Gene Symbols to Entrez GeneIDs}
\usage{
GeneSym2Entrez(DF, Species = NULL, GeneSym = "GeneName")
}
\arguments{
\item{DF}{a dataframe with a gene symbol column}

\item{Species}{One of "Human", "Mouse", "Rat", "Dog" or NULL.}

\item{GeneSym}{Column name of the gene symbol column to map}
}
\value{
DataFrame with Entrez GeneID column added
}
\description{
Maps Gene Symbols to Entrez GeneID using the org.Xx.eg.db packages
for Human, Mouse and Rat. Aggregates duplicate mappings (genesym mapped to 2
or more Entrez IDs) to avoid row expansion.
}
\details{
Note: currently aggregate mappings are lost due to exact matching.  Need
to implement a second pass for unmapped genesymbols to use partial matching.
}
\examples{
MyDataframe = GeneSym2Entrez (MyDataframe, Species = "Human", GeneName = "GeneSym")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Entrez,}
\keyword{GeneID}
\keyword{gene}
\keyword{symbol,}

