% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGE_Tools.R
\name{getLevel}
\alias{getLevel}
\title{Function  getLevel}
\usage{
getLevel(data)
}
\arguments{
\item{data}{Either an RangedSummarizedExperiment, SubsettableListOfArrays or
ContrastList object. A RangedSummarizedExperiment object}
}
\value{
Level of data (Gene, Transcript or Exon)
}
\description{
Return the Level (Gene or Transcript or Exon) of an RSE object
}
\examples{
MyLevel = getLevel(RSE)
MyLevel <- getLevel(MySLOA)
MyLevel <- getLevel(MyContrastData)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RangedSummarizedExperiment}
