% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotMDS.R
\name{ggplotMDS}
\alias{ggplotMDS}
\title{Function ggplotMDS}
\usage{
ggplotMDS(DGElist, top = Inf, labels = NULL, labelSize = 2,
  title = NULL, textColor = "blue2", hlineIntercept = NULL,
  vlineIntercept = NULL, reflineColor = "darkgoldenrod1",
  reflineSize = 0.5, baseFontSize = 12, themeStyle = "grey", ...)
}
\arguments{
\item{DGElist}{A DGEList object taken after running edgeR TMM normalization}

\item{top}{Number of most variant genes to include (default = Inf)}

\item{labels}{Text labels for the samples.  These should be short
abbreviations of the sample identifiers so that the plot is not too
cluttered.  If NULL, the column names from DGElist are used.}

\item{labelSize}{Size for the text labels in the plot (default = 2)}

\item{textColor}{Color for the text labels in the plot (default = "blue2")}

\item{hlineIntercept}{Y intercept of horizontal line (Default = Null)}

\item{vlineIntercept}{X intercept of vertical line (Default = Null)}

\item{reflineColor}{Color for the horizontal and vertiacl reference lines
(Default = "darkgoldenrod1")}

\item{reflineSize}{Thickness of the reference lines (Default = 0.5)}

\item{baseFontSize}{Base fontsize for the plot (Default = 12)}

\item{themeStyle}{One of "grey" or "bw" (Default = "grey")}

\item{...}{arguments from plotMDS.DGEList (see ?plotMDS.DGEList in package
edgeR)}
}
\value{
A list with two elements, the ggplot object and the mds object returned
   by the plotMDS function.
}
\description{
This is a wrapper around the plotMDS function that generates the plot with
ggplot2 instead of base graphics.
}
\details{
The underlying plotMDS function uses a default of top=500 to use the top 500
highest fold change genes for the analysis.  The method runs quick and I've
found situations where a larger value produced a more stable result.  Thus
I've change the default to top=Inf.  You should play with this parameter and
at least try a number that is close to the number of differential genes
detected in your data.
}
\examples{

     #Plot the first two dimensions using all genes
     MyMDS = ggplotMDS (MyDGEList)
     #display the plot and save a png file.
     printAndSave (MyMDS[[1]], file = "MyMDS.PNG", width=5, height=5)

     #plot the 2nd and 3rd dimensions using the top 1000 genes
     MyMDS = ggplotMDS (MyDGEList, dim.plot=c(2,3) ndim =3)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{compare}
\keyword{ggplot2}
\keyword{logratio}
\keyword{plot}
\keyword{scatterplot}

