% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOIplot.R
\name{GOIplot}
\alias{GOIplot}
\title{Provide a gene list of the genes of interest and return a plot of the LogFC data for each gene.
If >1 gene requested, the plot will be faceted which places an upper limit of around 25 or so
on the number of genes that can be displayed.}
\usage{
GOIplot(dgeObj, logRatioCol = "logFC", logIntCol = "AveExpr",
  pvalCol = "P.Value", pthreshold = 0.01, xlab, ylab, title,
  symbolSize = 2, symbolShape = 21, symbolColor = "black",
  symbolFill = "dodgerblue3", alpha = 0.5, zeroRefLine = TRUE,
  refLineThickness = 1, baseFontSize = 12, themeStyle = "grey", footnote,
  footnoteSize = 3, footnoteColor = "black", footnoteJust = 1)
}
\arguments{
\item{dgeObj}{A DGEobj workflow datastrucutre containing contrast (topTable) results (required)}

\item{logRatioCol}{name of the LogRatio column (Default = "logFC")}

\item{logIntCol}{name of the LogIntensity column (Default = "AveExpr")}

\item{pvalCol}{name of the pvalue or FDR column (Default = "P.Value")}

\item{xlab}{X axis label (default the LogIntensity column name)}

\item{ylab}{Y axis label (default the LogRatio column name)}

\item{title}{Plot title (optional)}

\item{symbolSize}{Size of symbols for Up, no change and Down. default = c(4,
3.99, 4); Note: you cannot choose the exact same size for all three.  But
you can use decimal values.}

\item{symbolShape}{Shape of the symbols for Up, no change and Down; Default =
c(21, 1, 21) (1 = open circle, 21 = fillable open circle); Note: you cannot
choose the same symbol shape for all three. See
\url{http://www.cookbook-r.com/Graphs/Shapes_and_line_types}}

\item{symbolColor}{c(Up, NoChange, Down); default = c("black", "grey25",
"grey0") See \url{http://research.stowers-institute.org/efg/R/Color/Chart}
Note: You cannot duplicate colors.}

\item{symbolFill}{Set the fill color for the symbols. Note only symbols 21-25
are fillable. This will have no effect on other symbols. Default =
c("red3", "grey25", "deepskyblue4") Note: You cannot duplicate colors.}

\item{alpha}{Controls the transparency of the plotted points (range: 0-1;
default = 0.7)}

\item{zeroRefLine}{Draw a referenceline at logFC=0 (Default = TRUE)}

\item{refLineThickness}{Size for the reference line (Default = 0.5)}

\item{baseFontSize}{The smallest size font in the figure in points. Default =
12}

\item{themeStyle}{"bw" or "grey" which correspond to theme_bw or theme_grey
respectively. Default = bw"}

\item{footnote}{optional string placed right justified at bottom of plot.}

\item{footnoteSize}{applies to footnote. (default = 3)}

\item{footnoteColor}{applies to footnote. (default = "grey90")}

\item{footnoteJust}{Value 0-1. 0 is left justified, 1 is right justified, 0.5 is centered. (default=1)}

\item{genelist}{List of ensembl gene IDs for the genes of interest (required)}

\item{plotType}{One of "point", "bar" or "line" (Default = "bar")}

\item{faceted}{Set to FALSE to return a list of ggplot objects, one per gene, instead of a single
faceted plot.}
}
\value{
ggplot object or list of ggplot objects
}
\description{
Provide a gene list of the genes of interest and return a plot of the LogFC data for each gene.
If >1 gene requested, the plot will be faceted which places an upper limit of around 25 or so
on the number of genes that can be displayed.
}
\examples{
   pending


}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{compare}
\keyword{ggplot2}
\keyword{logratio}
\keyword{plot}
\keyword{scatterplot}
