% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemePack.R
\name{greyTheme}
\alias{greyTheme}
\title{Function  greyTheme}
\usage{
greyTheme(base_size = 12, base_family = "", scale_legend = TRUE)
}
\arguments{
\item{base_size}{Size for the basefont in points (Default = 12)}

\item{base_family}{Set the font family}

\item{scale_legend}{Scales Legend.Text to 10/base_size lines (Default=TRUE)}
}
\value{
A ggplot theme that can be added to a plot object
}
\description{
A customization of the theme_grey default ggplot theme.  Use this instead of
theme_grey as the last layer on your plots.  This theme adds relative sizes
for most elements (title, axis labels and legends) so that it is easy to
adjust all font sizes without otherwise altering your plot. After applying
the greyTheme, use the baseFont function to alter just font sizes without
otherwise altering your plot. I also found that I didn't like the way legends
scaled when shifting from small fonts to larger fonts.  The legend tended to
grow unpleasingly large with larger fonts, especially when trying to locate
the legend inside the figure borders.  So the legend text is scaled by this
equation legendfont = 12/base_size.  You can turn this off with the
scaleLegend argument. Finally, I also use a darker color for the axis scale
labels (grey33 instead of grey50). I liked the idea of having the axis
numbering slightly lighter then the axis title but thought HW's choice of
grey50 for the axis numbering was too light.
}
\details{
Changes from the default theme_grey applied here are noted in source code comments.
}
\examples{
Myggplot = myggplot + greyTheme(24)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{fontsize}
\keyword{ggplot2}

