% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRT_heatmap.R
\name{JRT_heatmap}
\alias{JRT_heatmap}
\title{Function JRT_heatmap}
\usage{
JRT_heatmap(mat, myoptions = NULL)
}
\arguments{
\item{mat}{A numeric matrix of values for the heatmap}

\item{options}{A list of pheatmap options}
}
\value{
a pheatmap plot
}
\description{
JRT_heatmap is a wrapper around pheatmap.  Pheatmap has about 
40 settable command line arguments.  You might want to set
a dozen or so of those options different from the defaults
which makes for rather long lines.  But typically, you'll
set the same options for several heatmaps.  Enter JRT_heatmap.
I redefined the options as a list with sensible defaults
established.  For options common to many heatmaps, you only
need to set these options once at the top of your script.
Then you may need to set no more than the title for each
individual heatmap.
}
\details{
Another problem JRT_heatmap solve is uneven scales.  If you
scale is assymetric e.g. +6 to -4, then you center color (usually
white or black) will not be centered on zero with pheatmap.
JRT_heatmap adjusts the scale to insure that the desired center
color is centered on zero.

Peak at the source code for the full list of settable options.
}
\examples{
   #Draw a heatmap with all default values (2D cluster with trees)
		JRT_heatmap (myLogratios)

   #Add a title and change a few options
	 	#Here's a title that dynamically adds the number of genes displayed
   heatopts$main = paste("My Signature Heatmap\\n1D Clusters : ",
			nrow(LogRatios), " Genes", sep="")
	heatopts$cluster_cols = FALSE
	heatopts$clustering_distance_rows = "correlation"
	JRT_heatmap(LogRatios, heatopts)

	#Save the heatmap to a file instead of printing to the console
	heatopts$silent = TRUE
	heatopts$filename = "MyHeatmap.png"
	JRT_heatmap(LogRatios, heatopts)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{heatmap,}
\keyword{pheatmap}

