% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGE_Unit_Conversion.R
\name{Log2CPM_To_Log2FPKM}
\alias{Log2CPM_To_Log2FPKM}
\title{Function Log2CPM_To_Log2FPKM:  Convert Log2CPM to Log2FPKM}
\usage{
Log2CPM_To_Log2FPKM(log2cpm, exonLength)
}
\arguments{
\item{log2cpm}{A dataframe or matrix of Log2CPM values typically
from the voom  Elist ($E)}

\item{exonLength}{vector of total exon length from the design
(rowRanges) data.}
}
\value{
matrix of FPKM values
}
\description{
Convert Log2CPM from a voom elist (elist$E) to LOG2FPKM
to provide a Log2FPKM value to compare between
genes that is fully normalized and voom adjusted.
}
\examples{
MyFPKM = cpmToFpkm((2^log2cpm), exonlength)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{CPM;}
\keyword{FPKM}

