% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotMDS.R
\name{MDS_var_explained}
\alias{MDS_var_explained}
\title{Function MDS_var_explained}
\usage{
MDS_var_explained(
  mds,
  topN = 10,
  cumVarLimit = 0.9,
  barColor = "dodgerblue4",
  barFill = "dodgerblue3",
  barWidth = 0.65,
  barSize = 0.1,
  baseFontSize = 14
)
}
\arguments{
\item{mds}{A class MDS object from limma::plotMDS or a data matrix to analyze
(typically log2) (required)}

\item{topN}{The number of dimensions to plot (Default = 10)}

\item{cumVarLimit}{The maximum cumulative variance to plot. Range 0-1. (Default = 0.9)}

\item{barColor}{Default = "dodgerblue4"}

\item{barFill}{Default = "dodgerblue3"}

\item{barWidth}{range 0-1. (Default = 0.65)}

\item{barSize}{Thickness of the fill border (Default = 0.1)}

\item{baseFontSize}{Base fontsize for the plot (Default = 14)}
}
\value{
A list with two ggplots and the variance explained data.frame.
}
\description{
Takes a class MDS object from limma::plotMDS and generates two plots: 1)
fraction of variance for each dimension, 2) cumulative variance. By default,
it plots the first 10 dimensions or the first N dimensions totaling 90%. See
topN and cumVarLimit arguments to change the defaults.
}
\examples{

     #Plot the first two dimensions
     MyMDS = ggplotMDS (MyDGEList)
     MyMDS[[1]]  #the MDS plot

     #Then apply MDS_var_explained (the MDS object is MyMDS[[2]])
     varResults <- MDS_var_explained(MyMDS[[2]])
     varResults[[1]] #the Variance per dimension plot
     varResults[[2]] #the cumulative variance plot
     var_explained <- varResults[[3]]  #data used for plotting (unfiltered)

}
\author{
John Thompson, \email{jrt@thompsonclan.org}
}
\keyword{DGE,}
\keyword{MDS,}
\keyword{QC}
\keyword{RNA-Seq,}
