% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingDataHeatmap.R
\name{missingDataHeatmap}
\alias{missingDataHeatmap}
\title{Function missingDataHeatmap}
\usage{
missingDataHeatmap(data, filename, title = "", fontsize = 10,
  Rowv = NA, Colv = NA, color = c("firebrick2", "white",
  "dodgerblue3"), ...)
}
\arguments{
\item{data}{A dataframe with rownames to examine}

\item{filename}{Optionally provide a filename to save the graphic to a file instead of
sending to the console}

\item{title}{Optional text title for the plot}

\item{fontsize}{base fontsize for the plot (Default = 10)}

\item{Rowv}{clustering specification for rows (Default = NA) See ?NMF::heatmap for clustering options}

\item{Colv}{clustering specification for columns (Default = NA) See ?NMF::heatmap for clustering options}

\item{color}{Color map for the heatmap (Default = c("firebrick2", "white", "dodgerblue3")) See ?NMF:aheatmap for color options.}

\item{...}{Other arguments to pass to NMF::aheatmap}
}
\value{
A heatmap generated by aheatmap
}
\description{
Takes a dataframe as input and highlights where data is missing with a heatmap.  No clustering
is performed by default.
}
\examples{

   #make up some data with one row of NAs
   df <- data.frame(Doubles=c(1:10, NA, 12:20),
                    Characters=c(rep("A",10), NA, rep("B",9)),
                    stringsAsFactors=FALSE)
   rownames(df) <- paste("A", rownames(df), sep="")

   #Draw a heatmap of missing data
   missingDataHeatmap (df)

   #Save the heatmap to a file
   missingDataHeatmap (df, filename="./output/MyHeatmap.png")

   #apply some extra aheatmap options
   missingDataHeatmap(df, title="Test Title", legend=FALSE, sub="Red indicates missing data")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{data}
\keyword{heatmap}
\keyword{missing}
