% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemePack.R
\name{NoLegend}
\alias{NoLegend}
\title{Function  NoLegend}
\usage{
NoLegend()
}
\value{
A ggplot theme that can be added to a plot object
}
\description{
A simple function to spare me from looking up the syntax everytime
I want to remove the legend.
}
\examples{
Myggplot = myggplot + NoLegend()

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{fontsize}
\keyword{ggplot2}

