% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemePack.R
\name{NoXlab}
\alias{NoXlab}
\title{Function  NoXlab}
\usage{
NoXlab()
}
\value{
A ggplot theme that can be added to a plot object
}
\description{
A simple function to spare me from looking up the syntax everytime
I want to remove the X axis label.
}
\examples{
Myggplot = myggplot + NoXlab()

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{fontsize}
\keyword{ggplot2}
