% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemePack.R
\name{NoYGrid}
\alias{NoYGrid}
\title{Function  NoYGrid}
\usage{
NoYGrid()
}
\value{
A ggplot theme that can be added to a plot object
}
\description{
A simple function to spare me from looking up the syntax everytime
I want to remove the Y grid.
}
\examples{
Myggplot = myggplot + NoYGrid()

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{fontsize}
\keyword{ggplot2}
