% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsPlot.R
\name{obsPlot}
\alias{obsPlot}
\title{Deluxe obsPlot}
\usage{
obsPlot(data, block, blockOrder, obsNames, sampNames, plotBy = "Gene",
  valType = "Log2CPM", boxLayer = TRUE, violinLayer = FALSE,
  pointLayer = TRUE, meanLayer = TRUE, xlab = NULL, ylab = NULL,
  title = NULL, boxColor = "grey30", boxFill = "deepskyblue3",
  boxAlpha = 0.5, boxNotch = FALSE, boxNotchWidth = 0.8,
  violinColor = "grey30", violinFill = "goldenrod1", ViolinAlpha = 0.5,
  pointColor = "grey30", pointFill = "dodgerblue4", pointShape = 21,
  pointAlpha = 1, pointSize = 2, pointJitter = 0, meanColor = "red2",
  meanFill = "goldenrod1", meanShape = 22, meanAlpha = 0.7,
  meanSize = 3, legenPosition = "right", baseFontSize = 12,
  themeStyle = "grey", facet = TRUE, facetCol = NULL, xAngle = 30,
  scales = "free_y", returnPlotDat = FALSE)
}
\arguments{
\item{data}{matrix or dataframe of whatever data you want to plot with
samples in columns and observations (genes) in rows.}

\item{block}{Determines which samples belong to the same group.  Must be same
length as ncol(data).  Assign the same value to each member of a group.
Note that the block values are used to label the X Axis to identify the
groups.  Thus short pneumonic labels are useful here.}

\item{blockOrder}{Use this to explicitly set the display order of the blocks
in the plots.  blockOrder should contain unique(block) in the order you
wish to items to be arranged along the X-axis.  (optional)}

\item{obsNames}{list of row (observation) names to use instead of actual
rownames. (Default = NULL)}

\item{sampNames}{list of column (sample) names to use instead of actual
colnames. (Default = NULL)}

\item{plotBy}{Name for the type of observation being plotted (Default = "Gene")}

\item{valType}{name for the Value being plotted.  (Default = "Log2CPM")}

\item{boxLayer}{Adds a boxplot layer (Default = TRUE)}

\item{violinLayer}{Adds a violin layer (Default = FALSE)}

\item{pointLayer}{Adds a point layer (Default = True)}

\item{meanLayer}{Adds a mean layer (Default = True)}

\item{xlab}{X axis label (defaults to "Block")}

\item{ylab}{Y axis label (defaults to valType)}

\item{title}{Plot title (optional; Defaults = NULL)}

\item{boxColor}{Color for the boxplot layer (Default = "grey30")}

\item{boxFill}{Fill Color for the boxplot layer (Default = "deepskyblue3")}

\item{boxAlpha}{Transparency for the box layer (Default = 0.5)}

\item{boxNotch}{turn on/off confidence interval notches on boxplots (Default = FALSE)}

\item{boxNotchWidth}{Set the width of boxnotches (0-1) (Default = 0.8)}

\item{violinColor}{Color for the violin layer (Default = "grey30")}

\item{violinFill}{Fill Color for the violin (Default = "yellow")}

\item{pointColor}{Color for the point layer (Default = "grey30")}

\item{pointFill}{Fill color for the point layer (Default = "dodgerblue4")}

\item{pointShape}{Shape for the point layer (Default = 21; fillable circle)}

\item{pointAlpha}{Transparency for the box layer (Default = 1)}

\item{pointSize}{Size of the points (Default = 4)}

\item{pointJitter}{Amount to jitter the points (Default = 0) Try 0.2 if you
have alot of points.}

\item{meanColor}{Color for the point layer (Default = "red2")}

\item{meanFill}{Fill color for the point layer (Default = "goldenrod1")}

\item{meanShape}{Shape for the point layer (Default = 21; fillable circle)}

\item{meanAlpha}{Transparency for the box layer (Default = 0.7)}

\item{meanSize}{Size of the points (Default = 6)}

\item{baseFontSize}{The smallest size font in the figure in points. (Default =
12)}

\item{themeStyle}{"bw" or "grey" which correspond to theme_bw or theme_grey
respectively. Default = bw"}

\item{facet}{Specifies whether to facet (TRUE) or print individual plots
(FALSE)  (Default = TRUE)}

\item{xAngle}{Angle to set the sample labels on the Xaxis. (Default =  30; Range = 0-90)}

\item{scales}{Specify same scales or independent scales for each subplot (Default = "free_y";
Allowed values: "fixed", "free_x", "free_y", "free")}

\item{returnPlotDat}{Returns the dataframe used for the plot as a list member (default=FALSE)}

\item{violinAlpha}{Transparency for the box layer (Default = 0.5)}

\item{facetRow}{Explicitly set the number of Rows for the facet plot.
Default behavior will automatically set the columns. (Default = NULL)}
}
\value{
ggplot If Facet=TRUE (default) returns a facetted ggplot object. If
  facet=FALSE or returnPlotDat=TRUE, returns a list of ggplot objects indexed
  by observation (gene) names. If returnPlotDat=TRUE, the last element
  of the list is the dataframe used to generate the plot.
}
\description{
Provides a summary plot for each observation (gene), showing data for each
experiment group. The plot can optionally include one or more of the
following layers: boxplot, violin plot, individual points and/or mean of all
points.  The layers are built up in the order listed with user settable
transparency, colors etc.  By default, the Boxplot, Point and Mean layers are
active. Also, by default, the plots are faceted.  Facet plot can be turned
off to return a list of individual ggplot graphics for each gene.
}
\details{
Input is a dataframe or matrix of observations (rows; usually genes) by
Samples (columns) and requires rownames to identify the observations (genes).
A Block vector is required to define which Samples (columns) belong to the
same group and generate a summary plot by group with a separate plot for
each observation (gene).

Rownames are required and all the data columns should be numeric.  Each
observation (gene) generates a separate plot, so you should pass a smallish
list of genes unless you want alot of output.  By default the plot is
faceted. You should turn facet off if you have more than ~25 genes to plot.
Colnames and Rownames will be used to label the plots by default but custom
col/row names can also be supplied as separate vectors.
}
\examples{
   Simple faceted plot with custom title

   #get Log2CPM from an DGEobj object
   dgeObj = readRDS("MyDGEobj.RDS")
   dgelist <- getItem(dgeObj, "DGEList")
   Log2CPM <- cpm(dgelist, log=TRUE)
   genes = Log2CPM[1:12,] #first dozen genes
   #define six treatment groups in triplicate
   MyBlock = c(1,1,1,2,2,2,3,3,3,4,4,4,5,5,5,6,6,6)
   MyPlot = obsPlot(genes, MyBlock, title = "Plot Title")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{boxplot}
\keyword{ggplot2}
\keyword{logratio}
\keyword{violinplot}
