% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsoftRefGeneID2GeneSym.R
\name{OmicsoftRefGeneID2GeneSym}
\alias{OmicsoftRefGeneID2GeneSym}
\title{Function OmicsoftRefGeneID2GeneSym:  Convert Omicsoft RefGene GeneIDs to GeheSym}
\usage{
OmicsoftRefGeneID2GeneSym(GeneID)
}
\arguments{
\item{GeneID}{a vector of Omicsoft GeneIDs}
}
\value{
A vector with GeneSymbols
}
\description{
Maps Omicsoft RegGene IDs (e.g. NOC2L_1_-_1) to official gene symbol (e.g. NOC2L).
Omicsoft appends chr information to the genesymbol to make them unique.
This procedure simply right trims everything past the underscore
}
\examples{
#Senario GeneID is the rowname.  Add a GeneSym column
MyDF$MyGeneSym = OmicsoftRefGeneID2GeneSym (rownames(MyDF))

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Entrez,}
\keyword{GeneID}
\keyword{gene}
\keyword{symbol,}

