% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsoftToDgeObj.R
\name{OmicsoftToDgeObj}
\alias{OmicsoftToDgeObj}
\title{Function  OmicsoftToDgeObj}
\usage{
OmicsoftToDgeObj(counts = "RNA-Seq.Count.Table.txt",
  seqAnnotation = "RNA-Seq.Count.Annotation.txt",
  design = "RNA-Seq.Design.txt", level = "gene", path = ".", customAttr)
}
\arguments{
\item{counts}{A matrix or dataframe of R gene by C samples (required)
[Default = "RNA-Seq.Count.Table.txt"]}

\item{seqAnnotation}{Gene, isoform or exon level (row) annotation (required)
[Default = "RNA-Seq.Count.Annotation.txt"]}

\item{design}{Sample annotation with expt factors and other sample-associated
data (required) [Default = RNA-Seq.Design.txt"]}

\item{level}{One of "gene", "isoform", "exon" (required) [Default = "gene"]}

\item{path}{File path for the three data files (Default = "./")}

\item{customAttr}{A named list of custom attributes to attach to the DGEobj; 
Optional but highly encouraged;  suggestions: list(PID = "20170101-0001",
XpressID = "123456", Genome = "Mouse.B38", GeneMobel = "Ensembl.R84")}
}
\value{
A DGEobj
}
\description{
Build a DGEobj from Omicsoft output or tabbed text files.
}
\details{
User provides separate tab-delimited text files for row (gene) annotation,
column (sample) anotation and "assays" (matrices of RxC e.g. counts, FPKM, etc).

Two global variables hold the names of the Omicsoft data files; .geneData 
and .isoformData.

Data Requirements:

Assay data should have rownames (sequence ids) and column names (sample IDs).
Sequence annotation should also have the same rownames as assays.

If possible, the sequence annotation should include chromosome position data (chr,
start, end, strand).
}
\examples{

#Defaults set for an omicsoft dataset:
MyDgeObj <- OmicsoftToDgeObj (customAttr=list(
                                   PID = "20170101-0001",
                                   XpressID = "123456", 
                                   Genome = "Mouse.B38", 
                                   GeneMobel = "Ensembl.R84")

#I have data files from somewhere else                                   
MyDgeObj = OmicsoftToDgeObj (MyCounts, MyGeneAnnotation, MyDesign,
                                   level = "gene",
                                   customAttr = list(
                                      PID = "20170101-0001",
                                      XpressID = "123456", 
                                      Genome = "Mouse.B38", 
                                      GeneMobel = "Ensembl.R84")
                             )

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEObj,}
\keyword{Data}
\keyword{Omicsoft,}
\keyword{RNA-Seq,}
\keyword{loading}
