% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDispersion.R
\name{plotDispersion}
\alias{plotDispersion}
\title{Function  plotDispersion}
\usage{
plotDispersion(MyCounts, MyFormula = NULL, MyDesign = NULL,
  showPlot = TRUE, savePlot = TRUE)
}
\arguments{
\item{MyCounts}{A martrix of count data with geneID as rownames.}

\item{MyFormula}{a text string specifying the model formula}

\item{MyDesign}{Experiment design dataframe with columns referenced by the forumula}

\item{showPlot}{Set TRUE to print the plot}

\item{savePlot}{Set TRUE to save plot in a .PNG file}
}
\value{
nothing
}
\description{
Warning: Deprecated. Kept for now for backward compatibility but will be
removed in a later version. Replaced by DGE.Tools::plotDisp. See ?plotDisp
for details on the replacement function.
}
\details{
Runs an edgeR dispersion plot for RNA-Seq QC purposes.
}
\examples{
   plotDispersion (MyCounts, MyDesignMatrix, showPlot=TRUE, savePlot=TRUE)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq,}
\keyword{dispersion,}
\keyword{plot}

