% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNorm.R
\name{plotNorm}
\alias{plotNorm}
\title{Function plotNorm}
\usage{
plotNorm(dat, plotType = "box", normalize = "tmm", baseFont = 12)
}
\arguments{
\item{dat}{A DGEobj or counts matrix}

\item{plotType}{One of "box" or "density" (default="box")}

\item{normalize}{Default = "TMM" and invokes TMM normalization. Other allowed
values are: "RLE","upperquartile", "none". Invokes edgeR::calcNormFactors for
normalization.}

\item{baseFontSize}{Passed on as base font size in ggplot theme. (default=12)}
}
\value{
A  facetted ggplot plot showing before/after normalization
}
\description{
Takes a DGEobj or counts matrix as input. Returns ggplot object containing
a faceted plot of before/after normalization. Either a box plot or density plot
type can be chosen.
}
\examples{
   MyggPlot <- plotNorm(MyDgeObj, plotType="box")
   MyggPlot <- plotNorm(counts, plotType="density")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
