% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPValDist.R
\name{plotPvalDist}
\alias{plotPvalDist}
\title{Function  plotPvalDist}
\usage{
plotPvalDist(P.Val, Facet = TRUE, savePlot = TRUE, fileNames = NULL)
}
\arguments{
\item{P.Val}{a matrix or dataframe of numeric data; col=samples}

\item{Facet}{TRUE adds the samplename as a title bar on each facet}

\item{savePlot}{TRUE saves the plot(s) to .PNG files}

\item{fileNames}{Only relevant when Facet = FALSE. Default behavior takes png file
names from colnames.  Supply the FileNames parameter with a list of filenames
to override the default naming.  Must be same length as ncol(P.Val)}
}
\value{
A ggplot2 object or a list of plots
}
\description{
Generate a facet plot (or optionally individual plots) from a dataframe of
numbers.  Intended to perform histogram analysis of pvalue distributions.
But should be useful for any dataframe of numeric columns.
}
\examples{
MyDataframe = plotPvalDist (MyPvalDF, Facet = TRUE, SavePlot = TRUE)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Entrez,}
\keyword{GeneID}
\keyword{gene}
\keyword{symbol,}

