% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printAndSave.R
\name{printWithFootnote}
\alias{printWithFootnote}
\title{Function  printWithFootnote}
\usage{
printWithFootnote(plotObject, footnote, fontface = "plain", fontsize = 10,
  hjust = -0.1)
}
\arguments{
\item{plotObject}{A ggplot2 plotobject}

\item{filename}{A path/filename for the graphic}

\item{width}{Graphic width in inches (default = 7)}

\item{height}{Graphic height in inches (default = 5)}

\item{units}{Units for height and width ("in"|"cm"|"mm") (Default = "in")}

\item{scale}{Multiplicative scaling factor (Default = 1)}

\item{res}{Resolution in ppi (default=300)}

\item{printFontSize}{Base font size for the graphic on the console/knitr (default=12)}

\item{saveFontSize}{Base font size for the graphic file (default=24)}

\item{scaleLegend}{Scale the legend smaller if font > 14  (Default = TRUE)}

\item{printPlot}{Print to console if TRUE (Default=TRUE)}

\item{savePlot}{Print to file if TRUE (Default = TRUE)}
}
\value{
The print object
}
\description{
Print a ggplot2 object to the console/knitr report adding footnote text
under the plot.  Use when you want the footnot to be underneath the plot
labels.  Only prints the footnote once on a facetted plot.
}
\examples{
    #Write to the console or knitr report
    printWithFootnote(Myggplot, footnote = "Footnote Text")

    #Capture to a file
    png ("myplot.png", width=5, height=4, unit="in")
    printWithFootnote(Myggplot, footnote = "Footnote Text")
    invisible(dev.off())

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{bmp,}
\keyword{ggplot2,}
\keyword{jpeg,}
\keyword{pdf}
\keyword{png,}
\keyword{tiff,}
