% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCplots.R
\name{QCplots}
\alias{QCplots}
\title{Function QCplots}
\usage{
QCplots(qcdata, metricNames, plotType = "bar",
  barColor = "dodgerblue4", barFill = "dodgerblue3", barSize = 0.1,
  barAlpha = 1, barWidth = 0.9, pointColor = "grey30",
  pointFill = "dodgerblue4", pointShape = 21, pointAlpha = 1,
  pointSize = 2, lineColor = "dodgerblue4", lineSize = 1,
  lineType = "solid", lineFit = "loess", xAngle = 45,
  debug = FALSE)
}
\arguments{
\item{qcdata}{A dataframe or tibble with metric names in the first column and
samples in columns 2-n.  Each row is a different QC metric. This matches
the Omicsoft RNA-Seq.QCMetrics.Table.txt format.}

\item{metricNames}{A list of metrics to plot.  Values must exist in column 1
of the data frame.}

\item{plotType}{One of "bar", "point", "pointline".  If you want a different
plottype for each metric, pass a list of plotTypes with length equal to
length(metricNames)}

\item{barColor}{Color for the bar outline (default = "dodgerblue4")}

\item{barFill}{Color for the bar area (default = "dodgerblue3")}

\item{barSize}{set the bar size (thickness of each bar perimeter; default =
0.1)}

\item{barAlpha}{Transparency for the bar layer (Default = 1)}

\item{barWidth}{set the bar width (Default = 0.8)}

\item{pointColor}{Color for the point layer (Default = "grey30")}

\item{pointFill}{Fill color for the point layer (Default = "dodgerblue4")}

\item{pointShape}{Shape for the point layer (Default = 21; fillable circle)}

\item{pointAlpha}{Transparency for the box layer (Default = 1)}

\item{pointSize}{Size of the points (Default = 4)}

\item{lineColor}{Color of the line fit (Default = "dodgerblue4")}

\item{lineSize}{Size of the line fit (Default = 1)}

\item{lineType}{One of c("solid", "dashed", "dotted", "dotdash", "longdash",
"twodash"). (Default = "solid")}

\item{lineFit}{Type of fit to use.  One of c("auto", "lm", "glm", "gam",
"loess"). (Default = "loess")}

\item{xAngle}{Angle to set the sample labels on the Xaxis (Default =  45;
Range = 0-90)}

\item{debug}{Turn on debug mode (default = FALSE)}
}
\value{
ggplot object if one plot is specified.  A list of ggplot objects if 2 or more metrics specified
}
\description{
This function takes a dataframe with metric names in col1 and samples in col2-n.  Each row is a different metric.
It then generates the qc pltos as defined by other arguments.
}
\examples{


}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{barplot}
\keyword{confidence}
\keyword{contrasts}
\keyword{gplot2}
\keyword{intervals}
\keyword{lineplot}
\keyword{logratio}
