% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq_calc.R
\name{rsq_calc}
\alias{rsq_calc}
\title{Function  rsq_calc}
\usage{
rsq_calc(x, fit)
}
\arguments{
\item{x}{A normalized log2cpm matrix}

\item{fit}{A MArrayLM object from lmFit}
}
\value{
A vector of R-squared values for each gene fit.
}
\description{
Takes a Log2CPM numeric matrix and MArrayLM fit object from lmFit and calculates R-squared for each gene fit.
}
\examples{

  rsq  = rsq_calc (log2cpm, fitObject)

}
\references{
https://support.bioconductor.org/p/17431/  (Mark Robinson)
}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{R-squared}
\keyword{RNA-Seq;}
