% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runContrasts.R
\name{runContrasts}
\alias{runContrasts}
\title{Function  runContrasts}
\usage{
runContrasts(sloa, ContrastList, runTopTable = TRUE, runTopTreat = FALSE,
  FoldChangeThreshold = 1.5, runEBayes = TRUE, robust = TRUE,
  qvalues = FALSE, PvalueThreshold = 0.01, FDRthreshold = 0.1,
  proportion = 0.01, SigTable = NULL, SigTablePNG = NULL)
}
\arguments{
\item{sloa}{An SLOA object containing a Fit object}

\item{ContrastList}{A named list of contrasts}

\item{runTopTable}{runs topTable on the specified contrasts (Default = TRUE)}

\item{runTopTreat}{runs topTreat on the specified contrasts (Default = FALSE)}

\item{FoldChangeThreshold}{Only applies to TopTreat (Default = 1.5)}

\item{runEBayes}{Runs eBayes after lmfit; default = TRUE}

\item{robust}{eBayes robust option (default = TRUE)}

\item{qvalues}{Deprecated and ignored.  Qvalue functionality has been removed from
runContrasts and relegated to a separate function see ?runQvalues.#'}

\item{PvalueThreshold}{Default = 0.01}

\item{FDRthreshold}{Default = 0.1}

\item{proportion}{Proportion of genes expected to be differentially expressed
(used by eBayes) (Default = 0.01)}

\item{SigTable}{Deprecated and ignored: replaced with summarizeSigCounts function}

\item{SigTablePNG}{Deprecated and ignored: replaced with summarizeSigCounts function}
}
\value{
A list containing contrast dataframes and other associated data.frames.
}
\description{
Take an SLOA object and a named list of contrasts to perform. The SLOA must
contain a Fit object and a DesignMatrix. Returns a list of data objects
including topTable and topTreat output, ContrastMatrix, Fit.Contrasts, gene
and sample annotation (rowData and colData).
}
\details{
The contrasts are composed of column names from the DesignMatrix of the SLOA
and put into list form to start the analysis. Each contrast is named to give
it a short pneumonic name.

Example ContrastList \cr

ContrastList = list( \cr
   T1 = "treatment1 - control", \cr
   T2 = "treatment2 - control" \cr
) \cr

where treatment1, treatment2 and control are columns in the DesignMatrix

The returned ContrastAnalysis list contains the followin objects:
\itemize{
   \item{"ContrastMatrix"} {a matrix}
   \item{"Fit.Contrasts"} {a Fit object}
   \item{"TopTableList"} {a List of dataframes}
   \item{"TopTreatList"} {a List of dataframes}
   \item{"SigCountsTable"} {a dataframe summarizing signature gene counts}
   \item{"workflowRecord"} {List of reproducible documentation}
   \item{"Level"} {Type of data; one of Gene, Transcript or Exon}
}

See ?extractCol for a convenient means to extract a named column from all
the contrasts in a TopTableList or TopTreatList.
}
\examples{
#run defaults
MyContsratList  = runContrasts (MySLOA, ConstrastList)
MyContsratList  = runContrasts (MySLOA, ConstrastList, runTopTable = TRUE
       ConstrastType = "TopTreat", FoldChangeThreshold = 1.25)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{SLOA;}
\keyword{contrasts}

