% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdgeRNorm.R
\name{runEdgeRNorm}
\alias{runEdgeRNorm}
\title{Function  runEdgeRNorm (edgeR normlization )}
\usage{
runEdgeRNorm(RSE, Formula, NormMethod = "TMM", Design = NULL,
  plotFile = "Norm.Factors.PNG", plotLabels = NULL,
  estimateDispersion = TRUE)
}
\arguments{
\item{RSE}{An RSE object e.g as built by Build_RSE function}

\item{Formula}{a text string specifying the model formula. The formula must
use column names from the sample annotation in the RSE.DGE}

\item{NormMethod}{One of "TMM", "RLE", "upperquartile" or "none". Default = "TMM"}

\item{Design}{Used to override the design (colData) embedded in the RSE. Defaults to NULL}

\item{plotFile}{Enable a Barplot of the norm.factors produced (Default = "Norm.Factors.PNG")
Set to NULL to disable the plot.}

\item{plotLabels}{Sample labels for the plot. Length must equal the number of
samples. (Default = NULL;  sample number will be displayed)}

\item{estimateDispersion}{If true, run edgeR estimateDisp function and adds
dispersion values to the SLOA output.  (Default = True)  This takes a
minute or two ro run and you may want to turn it off when you re-running a
script during development or maybe better, use the cache=TRUE chunk opion
in knitr so save time rerunning blocks that haven't changed.}
}
\value{
SLOA a SubsettableListOfArrays containing data ready for DGE analysis
}
\description{
Returns a SubsettableListOfArrays (SLOA) containing the Fit and associated
output from the standard RNA-Seq using edgeR TMM normalization
followed by limma voom and lmfit.  The SLOA contains the Fit object
and various associated data  use "print(SLOA)" to examine the contents.
}
\examples{
MyFormula <- "~ 0 + DiseaseTreatment"
MySLOA <- runEdgeRNorm (RSE, MyFormula)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Entrez,}
\keyword{GeneID}
\keyword{gene}
\keyword{symbol,}

