% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIHW.R
\name{runIHW}
\alias{runIHW}
\title{Function  runIHW}
\usage{
runIHW(contrastList, alpha = 0.1, FDRthreshold = 0.1, ...)
}
\arguments{
\item{contrastList}{A list object produced by runContrasts.  The list must
contain, either or both of TopTableList and TopTreatList items as provided
by the runContrasts function.}

\item{alpha}{The alpha value is the FDR level you wish to interogate (range 0-1; default = 0.1)}

\item{FDRthreshold}{Value used to estimate proportion of differential genes for each contrast.}

\item{...}{other arguments are passed directly to the ihw function (see ?ihw).}
}
\value{
The contrastList is returned.  The TopTableList and TopTreatList items will
contain additional columns added by the ihw analysis and prefixed with "ihw."
The IHW dataframe of results will also be appended to the contrast list.
}
\description{
This is a wrapper around the independent hypothesis weighting package that
takes a contrast list from runContrasts and applies Independent Hypothesis
Weighting (IHW) to the TopTableList within the contrast list.  IHW is a
method developed by N. Ignatiadis (http://dx.doi.org/10.1101/034330) to weight
FDR values based on a covariate (AveExpr in this case).
}
\details{
The IHW FDR values are added as additional columns to the topTable dataframes.
It is designed to work on dataframes created by topTable and containe in the
"contrast list" data structure produced by runContrasts.

Note that it is impractical to run IHW on a list of genes less than ~5000 because
it breaks the data into bins of 1500 genes for the analysis if bins = 1, IHW converges
on the BH FDR value and there's no point.  So we run IHW on the complete set
of detected genes from topTable results (not topTreat).
}
\examples{
MyContrasts = runIHW(MyContrasts)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{bmp,}
\keyword{ggplot2,}
\keyword{jpeg,}
\keyword{pdf}
\keyword{png,}
\keyword{tiff,}
