% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemePack.R
\name{setLegendPosition}
\alias{setLegendPosition}
\title{Function  setLegendPosition}
\usage{
setLegendPosition(ggplot, legendPosition = "right", themeStyle = "grey")
}
\arguments{
\item{ggplot}{A ggplot plot object}

\item{legendPosition}{One of top, bottom, right, left, ne, se, nw, sw.  The
first four place it outside the figure. The last four place it inside the
figure. (Default = "right")}

\item{themeStyle}{one of "grey" or "bw". (Default = "grey")}
}
\value{
A ggplot object with the legend adjusted as specified
}
\description{
Move the legend around outside (right, left, top, bottom) or inside corner
(ne, se, nw, sw) of your plot.  Respects the bw or grey theme style.  You
could also use this to get a grey legend on a bw theme or vice versa.
}
\details{
This should be called after setting one of greyTheme, bwTheme,
theme_grey or theme_bw as those theme reset the legend attributes.
}
\examples{
Myggplot = myggplot + NoXGrid()

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{fontsize}
\keyword{ggplot2}

