% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeSigCounts.R
\name{summarizeSigCounts}
\alias{summarizeSigCounts}
\title{Function  summarizeSigCounts}
\usage{
summarizeSigCounts(
  ContrastList,
  columns = c("P.Value", "adj.P.Val", "Qvalue", "qvalue.lfdr", "ihw.adj_pvalue"),
  sigThresholds = c(0.01, 0.1, 0.1, 0.1, 0.1),
  fcThreshold = 0
)
}
\arguments{
\item{ContrastList}{A list of topTable dataframes}

\item{columns}{Vector of column names to summarize from topTable dataframes
Default = c("P.Value", "adj.P.Val", "Qvalue", "qvalue.lfdr", "ihw.adj_pvalue")}

\item{sigThresholds}{Thresholds to use for each column specified in columns
Must be same length at columns argument.
Default = c(0.01, 0.1, 0.1, 0.1, 0.1)}

\item{fcThreshold}{fold-change threshold (absolute value, not logged)}
}
\value{
data.frame with one summary row per contrast
}
\description{
Takes a contrast list produced by runContrasts. Also, specify columns to
summarize and thresholds for each column.  Optionally, provide a fold change
threshold.  Queries the topTable results and returns a dataframe with the
summary results; gene counts that meet the specified conditions.
}
\details{
Specified column names that don't exist will be ignored. So normally, the
defaults cover all the pvalue and FDR related columns and you'll
only have to add a fcThreshold if you want one or modify the pvalue/fdr
thresholds if you want different thresholds from the defaults.
}
\examples{

#get a contrast list from a dgeObj
MyContrastList <- getType(dgeObj, "topTable")

#all default thresholds, no fold change threshold
MySigSummary <- summarizeSigCounts (MyContrastList)

#all defaults with a foldchange threshold
MySigSummary <- summarizeSigCounts (MyContrastList, fcThreshold=1.5)

#change the pvalue and fdr thresholds
MySigSummary <- summarizeSigCounts (MyContrastList,
    sigThresholds = c(0.05, 0.2, 0.2, 0.2, 0.2))

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGE.Tools,}
\keyword{counts}
\keyword{gene}
\keyword{signature}
\keyword{summary,}
