% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCounts.R
\name{tpm.classic}
\alias{tpm.classic}
\title{Function tpm.classic}
\usage{
tpm.classic(counts, genelength, collapse = FALSE)
}
\arguments{
\item{counts}{A numeric matrix of N genes x M Samples.  All columns
must be numeric.}

\item{collapse}{T/F determines whether to use rowMeans on the genelength matrix [Default = FALSE]}

\item{geneLength}{A vector or matrix of gene lengths.}
}
\value{
A matrix of TPM values
}
\description{
Takes a counts and genelength as input and converts to tpm units using the equation from 
[Harold Pimental](https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/).
}
\details{
Genelength can be a vector (length == nrow(counts) or a matrix (same dim as counts).
The genelength is used as is, or optionally collapsed to a vector by rowMeans.
}
\examples{
myTPM <- tpm.classic(mycounts, mygenelength) 

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq,}
\keyword{conversion}
\keyword{unit}
