% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCounts.R
\name{tpm.on.subset}
\alias{tpm.on.subset}
\title{Function tpm.on.subset}
\usage{
tpm.on.subset(dgeo, applyFilter = TRUE)
}
\arguments{
\item{dgeo}{A DGEobj data structure}

\item{applyFilter}{If TRUE, reduces to the filtered gene list. FALSE returns
all genes in the raw data. [Default = TRUE]}
}
\value{
A matrix in the new unit space
}
\description{
Use this to calculate TPM for a heavily subsetted DGEobj.  It calculates TPM
using the original data but returns a DGEobj with the subset.
}
\details{
Takes a DGEObj as input. Uses all data (pre-genefiltering; counts_orig) to
calculate TPM values. Then optionally filters any genes that were already
filtered out.

For GeneLength in the calculation, it takes geneData$ExonLength (for Omicsoft data)
or uses rowMeans(effectiveLength) for data derived from Xpress.

Internally it uses edgeR::fpkm to calculate fpkm and converts to tpm
using the formula provided  by [Harold Pimental](https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/).
}
\examples{
myTPM <- tpm(dgeobj)

}
\author{
John Thompson, \email{jrt@thompsonclan.org}
}
\keyword{RNA-Seq,}
\keyword{conversion}
\keyword{unit}
