% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGE_Tools.R
\name{tsmsg}
\alias{tsmsg}
\title{GeneData (list of lists)}
\usage{
tsmsg(...)
}
\description{
Defines the data files to import into a SummarizedExperiment.
Each file is described by a list with defines the fieldname, the type of data
and the text file containing the data.  The predefined GeneData item
is preconfigured for datafiles from the Omicsoft pipeline.  Modify GeneData
file values to use data from a different pipeline.
}
\details{
TranscriptData (list of lists)

Defines the data files to import into a SummarizedExperiment.
Each file is described by a list with defines the fieldname, the type of data
and the text file containing the data.  The predefined TranscriptData item
is preconfigured for datafiles from the Omicsoft pipeline.  Modify TranscriptData
file values to use data from a different pipeline.
}
