% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voomWorkflow.R
\name{voomWorkflow}
\alias{voomWorkflow}
\title{Function voomWorkflow}
\usage{
voomWorkflow(dgeObj, formula, projectName, designMatrixName,
  fracThreshold = 0.5, outputPath = "./", annotationFile,
  proteinCodingOnly = FALSE)
}
\arguments{
\item{dgeObj}{A class dgeObj with counts, gene annotation and sample annotation}

\item{formula}{A text representation of the formula you want to use}

\item{projectName}{This should be the project name from Xpress or Omicsoft}

\item{designMatrixName}{User defined name for the design matrix}

\item{fracThreshold}{Fraction of samples that must meet intensity thresholds to keep a gene (Default = 0.5)}

\item{outputPath}{Where to send output plots}

\item{annotationFile}{Text file of key=value pairs to populate DGEobj attributes (optional but highly advised)}

\item{proteinCodingOnly}{Set to TRUE to keep only protein coding genes (default = FALSE)}
}
\value{
A DGEobj with analysis results added
}
\description{
This function runs several steps in the RNA-Seq pipeline in one fell swoop.
It supports duplicateCorrelation if you provide a blocking vector.  It includes
low intensity filtering, filtering for protein coding genes and filters out zero
effective length genes (genes shorter than library size).  Then it runs TMM normalization,
voomWithQualityWeights, lmFit and eBayes.
}
\details{
To incorporate SVA analysis, you need to run SVA first and add the SVA variables to your design
table. Then you can proceed with this function.
}
\examples{

   MyDgeObj <- voomWorkflow(MyDgeObj,
                            formula = "~ 0 + ReplicateGroup",
                            projectName = "MyProjectName",
                            designMatrixName = "ReplicateGroup",
                            annotationFile = "MyProjectName.txt",
                            proteinCodingOnly = TRUE)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj,}
\keyword{RNA-Seq,}
\keyword{limma}
\keyword{voom}
