% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGE_Tools.R
\name{yrange}
\alias{yrange}
\title{Function  yrange}
\usage{
yrange(my.ggp)
}
\arguments{
\item{g}{A ggplot plot object (ggplot2 v3 or higher)}
}
\value{
A vector length 2
}
\description{
extract the Y upper and lower limits from a ggplot2 v3 plot object.
}
\examples{
myYrange = yrange (myggplot)

}
\author{
John Thompson, \email{rct@thompsonclan.org}
}
\keyword{ggplot,}
\keyword{limits}
\keyword{ranges,}
