% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGE_Tools.R
\name{eBayes_autoprop}
\alias{eBayes_autoprop}
\title{Function  eBayes_autoprop}
\usage{
eBayes_autoprop(..., prop.method = "lfdr")
}
\arguments{
\item{...}{See ?eBayes}

\item{prop.method}{defaults to "lfdr"}
}
\value{
SLOA a SubsettableListOfArrays containing data ready for DGE analysis
}
\description{
If you use the following function in place of eBayes for single contrasts,
you can actually use the B-statistic from the results of topTable.
The B-statistic is the log-odds of differential expression, so a zero
means a 50-50 chance of DE, positive means more likely DE than not,
and negative means more likely non-DE. (Note that this statistic uses
the natural logarithm, not base 10 or base 2 logs.) However, this
statistic is only accurate if you supply a prior estimate of the
fraction of genes that are DE. The default proportion is 1%,
which should give conservative values as long as at least 1% of
genes are DE.
}
\details{
The function below uses propTrueNull to determine this prior, which
means that the B-statistics should be reliable as long as propTrueNull
returns a reasonable-looking value, which you can verify by looking at
the p-value distribution.
}
\examples{
MyFit = eBayes_autoprop (MyFit)

}
\author{
Ryan Thompson, \email{rct@thompsonclan.org}
}
\keyword{DGE,}
\keyword{eBayes,}
\keyword{limma}
