% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCol.R
\name{extractCol}
\alias{extractCol}
\title{Function  extractCol}
\usage{
extractCol(dflist, ColName)
}
\arguments{
\item{dflist}{A list of data.frames which all have the same colnames.
The dataframes in the list should have rownames (geneIDs).}

\item{ColName}{The name of the data column to extract to a matrix}
}
\value{
A matrix containing the extracted columns
}
\description{
Take a list of dataframes where each dataframe has the same
column names (e.g. a list of topTable dataframes). Extract
the named column from each dataframe and return a matrix.
}
\examples{
MyPvalues  = ExtractCol (MySLOA$TopTableList, "P.Value")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{topTable}

