% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoformFrac.R
\name{isoformFrac}
\alias{isoformFrac}
\title{Function isoformFrac}
\usage{
isoformFrac(dgeObj)
}
\arguments{
\item{dgeObj}{A isoform level dgeObj created by function initDGEobj}

\item{normalize}{Default = TRUE and invokes TMM normalization. FALSE uses
TPM without TMM normalization.  Other allowed values are: "TMM","RLE","upperquartile". 
Invokes edgeR::calcNormFactors for normalization.}
}
\value{
A DGEobj with an isoform fraction dataframe added
}
\description{
Takes a DGEobj as input  and adds an item containing isoform fraction data.
}
\details{
Isoform Fraction is calculated using TPM data, as length normalized data
is required since different isoforms have different total exon lengths.  Isoform
fraction is calculated simply as the isoform TPM divided by the total TPM
for all isoforms of a given gene.

TPM is calculated from counts using the ExonLength column from the gene/isoform
data.  If normalization is specified, TMM or other selected normalization
is applied to CPM data (using edgeR functions) before converting to TMM.
}
\examples{
   MyDgeObj <- isoformFrac(MyDgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
