% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowIntFilter.R
\name{lowIntFilter}
\alias{lowIntFilter}
\title{Function  lowIntFilter}
\usage{
lowIntFilter(x, zfpkmThreshold, fpkThreshold, countThreshold,
  sampleFraction = 0.5, genelength = NULL, verbose = FALSE)
}
\arguments{
\item{x}{A counts matrix or data.frame OR a DGEobj with RNA-Seq (counts) data
(required).}

\item{zfpkmThreshold}{Genes below this threshold are removed (-3.0 is recommended)}

\item{fpkThreshold}{Genes below this threshold are removed (recommend 5).}

\item{countThreshold}{Genes below this threshold are removed (recommend 10).
Set to 0 to disable this filter.}

\item{sampleFraction}{The proportion of samples that must meet the thresholds
(Default = 0.5). Range >0 and <=1.}

\item{genelength}{Vector of genelengths for rows of x. Required for FPK and
zFPKM filters, unless x is a DGEobj.  In the DGEobj case, genelength is
retrieved from the DGEobj, unless supplied by the genelength argument.}

\item{verbose}{Prints some messages about the filtering process.}
}
\value{
Same class as input object with low intensity rows removed
}
\description{
Takes a DGEobj or counts matrix as input and applies a combination of low
intensity filters. raw count, zFPKM and/or FPK filters are supported.  A gene
must pass all active filters.  Not setting a threshold argument inactivates that threshold.
}
\examples{

  #simple count threshold
  myDgeObj  = lowIntFilter (myDgeObj, countThreshold=10)

  #count and zFPKM thresholds
  myDgeObj  = lowIntFilter (myDgeObj, countThreshold=10, zfpkmThreshold = -3.0)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq;}
\keyword{counts;}
\keyword{intensity}
\keyword{low}
