% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowIntFilter.R
\name{lowIntFilter}
\alias{lowIntFilter}
\title{Function  lowIntFilter}
\usage{
lowIntFilter(x, zfpkmThreshold = -3, fpkThreshold = 5,
  countThreshold = 10, sampleFraction = 0.5, genelength)
}
\arguments{
\item{x}{A counts matrix or data.frame OR DGEobj with RNA-Seq data
(required).}

\item{zfpkmThreshold}{Genes below this threshold are removed
(Default = -3.0).  Set to -Inf to disable
this filter.}

\item{fpkThreshold}{Genes below this threshold are removed (Default = 5).
Set to 0 to disable this filter.}

\item{countThreshold}{Genes below this threshold are removed (Default = 10 ).
Set to 0 to disable this filter.}

\item{sampleFraction}{The proportion of samples that must meet the thresholds
(Default = 0.5). Range >0 and <=1.}

\item{genelength}{Vector of genelength for rows of x (required unless x is a
DGEobj)}
}
\value{
Same class as input object with low intensity rows removed
}
\description{
Takes a DGEobj or counts matrix as input and applies a combination of low
intensity filters. zFPKM, FPK and raw Count filters are supported.  A gene
must pass all filters.
}
\examples{

  myDgeObj  = lowIntFilter (myDgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq;}
\keyword{counts;}
\keyword{intensity}
\keyword{low}
