% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowIntFilter.R
\name{lowIntFilter}
\alias{lowIntFilter}
\title{Function  lowIntFilter}
\usage{
lowIntFilter(x, xThreshold = -3, countThreshold = 10,
  sampleFraction = 0.5, genelength)
}
\arguments{
\item{x}{A counts matrix or data.frame OR DGEobj with RNA-Seq data.}

\item{countThreshold}{Default = 10 Minimum read counts to keep a gene}

\item{sampleFraction}{The proportion of samples that must meet the thresholds  Default = 0.5}

\item{genelength}{Vector of genelength for rows of x (required unless x is a DGEobj)}

\item{zThreshold}{Default = -3.0 Genes below this threshold are considered not detected.}
}
\value{
Same class as input object with low intensity rows removed
}
\description{
Take a named list of dataframes where each dataframe has the same
column names (e.g. a list of topTable dataframes). Extract
the named column from each dataframe and return a matrix.
}
\details{
The common use case for this is to provide a list of topTable
data frames and extract one column from each file to create
a matrix of LogRatios or Pvalues.

Technically, this should work as long as the requested colName is present
in each dataframe.  The default robust = TRUE should be used unless you
are absolutely certain each dataframe in the input list has the same row count
and row order.
}
\examples{

  myDgeObj  = lowIntFilter (myDgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq;}
\keyword{counts;}
\keyword{intensity}
\keyword{low}
