% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphDGEobj.R
\name{mapDGEobj}
\alias{mapDGEobj}
\title{Function mapDGEobj}
\usage{
mapDGEobj(dgeObj, directed = TRUE)
}
\arguments{
\item{DGEobj}{A workflow object of DGEobj class data.}
}
\value{
A class igraph network object.
}
\description{
Reads a DGEobj and produces a node pair file defining parent/child relationships between
data items in the DGEobj.  Plot them with Function graphDGEobj.
}
\examples{

  library(igraph)
  library(RColorBrewer)

  # Prepare an iGraph object for plotting
  mynet <- mapDGEobj(dgeObj)

  #define a layout
  lay.sug <- layout_with_sugiyama(mynet)

  #2D Plot
  plot(mynet,
       edge.arrow.size=.3,
       vertex.color="dodgerblue2",
       vertex.frame.color="dodgerblue4",
       layout=lay.sug$layout)

  #2D Plot; colorby the basetype attribute
  pal <- brewer.pal(length(unique(V(mynet)$basetype)), "Set1")
  myPallet <- pal[as.numeric(as.factor(vertex_attr(mynet, "basetype")))]

  plot(mynet,
       edge.arrow.size=.3,
       vertex.color = myPallet,
       vertex.label.family = "Helvetica",
       layout=lay.sug$layout
       )

  #2D Interactive plot
  plotHandle <- tkplot(net,
                       vertex.color=myPallet,
                       vertex.frame.color="dodgerblue4",
                       canvas.width = 800,
                       canvas.height = 800,
                       layout=lay.sug$layout)
  tk_close(plotHandle)

}
\author{
John Thompson, \email{jrt@thompsonclan.org}
}
\keyword{RNA-Seq,}
\keyword{conversion}
\keyword{unit}
