% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsPlot2.R
\name{obsPlot2}
\alias{obsPlot2}
\title{Function obsPlot2}
\usage{
obsPlot2(data, plotByCol, groupCol, valueCol,
  groupOrder = unique(as.character(data[groupCol, , drop = TRUE])),
  boxLayer = TRUE, violinLayer = FALSE, pointLayer = TRUE,
  meanLayer = TRUE, xlab = groupCol, ylab = valueCol, title,
  boxColor = "grey30", boxFill = "deepskyblue3", boxAlpha = 0.5,
  boxNotch = FALSE, boxNotchWidth = 0.8, violinColor = "grey30",
  violinFill = "goldenrod1", ViolinAlpha = 0.5,
  pointColor = "grey30", pointFill = "dodgerblue4", pointShape = 21,
  pointAlpha = 1, pointSize = 2, pointJitter = 0,
  meanColor = "red2", meanFill = "goldenrod1", meanShape = 22,
  meanAlpha = 0.7, meanSize = 3, legenPosition = "right",
  baseFontSize = 12, themeStyle = "grey", facet = TRUE, facetCol,
  xAngle = 30, scales = "free_y", debug = FALSE)
}
\arguments{
\item{data}{A tidy dataframe of intensity data.}

\item{plotByCol}{Define the column name to separate plots.}

\item{groupCol}{Define the column name to group boxplots by.}

\item{valueCol}{Define the column of values for plotting.}

\item{groupOrder}{Define the order for the groups in each plot.  Should
contain values in unique(data$group) listed in the order that you want the
groups to appear in the plot. (optional; default = unique(data[groupCol]))}

\item{boxLayer}{Adds a boxplot layer (Default = TRUE)}

\item{violinLayer}{Adds a violin layer (Default = FALSE)}

\item{pointLayer}{Adds a point layer (Default = True)}

\item{meanLayer}{Adds a mean layer (Default = True)}

\item{xlab}{X axis label (defaults to groupCol)}

\item{ylab}{Y axis label (defaults to valueCol)}

\item{title}{Plot title (optional)}

\item{boxColor}{Color for the boxplot layer (Default = "grey30")}

\item{boxFill}{Fill Color for the boxplot layer (Default = "deepskyblue3")}

\item{boxAlpha}{Transparency for the box layer (Default = 0.5)}

\item{boxNotch}{turn on/off confidence interval notches on boxplots (Default = FALSE)}

\item{boxNotchWidth}{Set the width of boxnotches (0-1) (Default = 0.8)}

\item{violinColor}{Color for the violin layer (Default = "grey30")}

\item{violinFill}{Fill Color for the violin (Default = "yellow")}

\item{pointColor}{Color for the point layer (Default = "grey30")}

\item{pointFill}{Fill color for the point layer (Default = "dodgerblue4")}

\item{pointShape}{Shape for the point layer (Default = 21; fillable circle)}

\item{pointAlpha}{Transparency for the box layer (Default = 1)}

\item{pointSize}{Size of the points (Default = 4)}

\item{pointJitter}{Amount to jitter the points (Default = 0) Try 0.2 if you
have alot of points.}

\item{meanColor}{Color for the mean layer (Default = "red2")}

\item{meanFill}{Fill color for the mean layer (Default = "goldenrod1")}

\item{meanShape}{Shape for the mean layer (Default = 21; fillable circle)}

\item{meanAlpha}{Transparency for the mean layer (Default = 0.7)}

\item{meanSize}{Size of the mean points (Default = 3)}

\item{baseFontSize}{The smallest size font in the figure in points. (Default =
12)}

\item{themeStyle}{"bw" or "grey" which correspond to theme_bw or theme_grey
respectively. Default = bw"}

\item{facet}{Specifies whether to facet (TRUE) or print individual plots
(FALSE)  (Default = TRUE)}

\item{xAngle}{Angle to set the sample labels on the Xaxis. (Default =  30; Range = 0-90)}

\item{scales}{Specify same scales or independent scales for each subplot (Default = "free_y";
Allowed values: "fixed", "free_x", "free_y", "free")}

\item{debug}{Turn on debug mode (default = FALSE)}

\item{violinAlpha}{Transparency for the box layer (Default = 0.5)}

\item{facetRow}{Explicitly set the number of Rows for the facet plot.
Default behavior will automatically set the columns. (Default = NULL)}
}
\value{
ggplot If Facet=TRUE (default) returns a facetted ggplot object. If
  facet=FALSE, returns a list of ggplot objects indexed
  by observation (gene) names.
}
\description{
Provides a summary plot for each observation (gene), showing data for each
experiment group. The plot can optionally include one or more of the
following layers: boxplot, violin plot, individual points and/or mean of all
points.  The layers are built up in the order listed with user settable
transparency, colors etc.  By default, the Boxplot, Point and Mean layers are
active. Also, by default, the plots are faceted.  Facet plot can be turned
off to return a list of individual ggplot graphics for each gene.
}
\details{
Input is a tidy dataframe of intensity data.
}
\examples{

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{boxplot}
\keyword{ggplot2}
\keyword{logratio}
\keyword{violinplot}
