% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDisp.R
\name{plotDisp}
\alias{plotDisp}
\title{Function  plotDispersion}
\usage{
plotDisp(x, designMatrix, symbolSize = 1, symbolShape = 1,
  symbolColor = "darkblue", symbolFill = "darkblue", symbolAlpha = 0.3,
  linefitSize = 1, linefitColor = "yellow", lineFit = NULL,
  lineType = 1, lineAlpha = 1, rugColor = NULL, rugAlpha = 0.02, ...)
}
\arguments{
\item{x}{A Counts matrix or DGEList (required)}

\item{designMatrix}{A design matrix created limma by model.matrix}

\item{symbolSize}{(Default=1)}

\item{symbolShape}{see
http://www.cookbook-r.com/Graphs/Shapes_and_line_types/ (Default = 1)}

\item{symbolColor}{Default = "darkblue"}

\item{symbolFill}{Default = "darkblue"}

\item{symbolAlpha}{Transparency for the points. Value from 0 to 1. Smaller
indicate more transparency (Default = 0.3)}

\item{linefitSize}{(Default = 1)}

\item{linefitColor}{(Default = "yellow")}

\item{lineFit}{(Default = NULL) Any type supported by geom_smooth (loess is
recommended). Defaults to NULL because enabling this
overloads my poor 32bit PC.}

\item{rugColor}{(Default = NULL)  Set to Null to disable rug plots.
Note: printing plots with rug plots can be slow (several minutes on my
meager 32bit pc). So I left this disabled by default.  Assign a valid
color to enable}

\item{rugAlpha}{Transparency for the rug plots (Default = 0.02)}

\item{...}{Extra parameters to pass to edgeR::estimateDisp}
}
\value{
a ggplot object
}
\description{
Creates an edgeR dispersion plot for RNA-Seq QC purposes.  Takes a SLOA object
for input after runEdgeRNorm has been executed.  The Biological coefficient
of variation (BCV) is displayed.  BCV is the sqrt of dispersion.
}
\examples{
   MyGgplot <- plotDisp (MyDGElist)
   MyGgplot <- plotDisp (MyDGEobj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{QC}
\keyword{RNA-Seq,}
\keyword{dispersion,}
\keyword{plot,}

