% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNorm.R
\name{plotNorm}
\alias{plotNorm}
\title{Function plotNorm}
\usage{
plotNorm(dat, plotType = "box", normalize = "tmm", baseFont = 12)
}
\arguments{
\item{dat}{A DGEobj or counts matrix}

\item{plotType}{One of "box" or "density" (default="box")}

\item{normalize}{Default = "TMM" and invokes TMM normalization. Other allowed
values are: "RLE","upperquartile", "none". Invokes edgeR::calcNormFactors for
normalization.}

\item{baseFontSize}{Passed on as base font size in ggplot theme. (default=12)}

\item{...}{additional arguments are passed to calcNormFactors.}
}
\value{
A  facetted ggplot plot showing before/after normalization
}
\description{
Takes a DGEobj (containing counts) or a counts matrix as input. Returns ggplot object containing
a faceted plot of log2CPM before/after normalization. Either a box plot or density plot
type can be chosen.
}
\details{
Normalization is performed by edgeR::calcNormFactors. Note TMM is specifically tailored to count-based
data.  Thus this function is only appropriate for count-based data.
}
\examples{
   MyggPlot <- plotNorm(MyDgeObj, plotType="box")
   MyggPlot <- plotNorm(counts, plotType="density")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
