% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printAndSave.R
\name{printAndSave}
\alias{printAndSave}
\title{Function  printAndSave}
\usage{
printAndSave(plotObject, filename, width = 7, height = 5, units = "in",
  res = 300, printFontSize = 12, saveFontSize = 24, scaleLegend = TRUE,
  printPlot = TRUE, savePlot = TRUE)
}
\arguments{
\item{plotObject}{A ggplot2 plotobject}

\item{filename}{A path/filename for the graphic}

\item{width}{Graphic width in inches (default = 7)}

\item{height}{Graphic height in inches (default = 5)}

\item{res}{Resolution in ppi (default=300)}

\item{printFontSize}{Base font size for the graphic on the console/knitr (default=12)}

\item{saveFontSize}{Base font size for the graphic file (default=24)}

\item{scaleLegend}{Scale the legend smaller if font > 14  (Default = TRUE)}

\item{printPlot}{Print to console if TRUE (Default=TRUE)}

\item{savePlot}{Print to file if TRUE (Default = TRUE)
(Default = TRUE)}
}
\value{
The print object
}
\description{
Print a ggplot2 object to the console/knitr report and save as a graphic file.
Defaults to larger fonts for the graphic file.  Graphic file type is taken from
the given filename and can be one of (png, bmp, tiff, jpg, pdf, svg).
}
\details{
I found I didn't like the default Legend sizes. I inserted code to increase the
size of the print/knitr legend and decrease the size of the legend in the saved
graphic file.  This works well for images of about 5-7 inches but may not scale
well to other image sizes.  Set scaleLegend to FALSE to get the default behavior
back or use legend.key.size in a theme to tweak to your own specifications.
}
\examples{
printAndSave (Myggplot, myfile.png) #all defaults
printAndSave (Myggplot, myfile.png, width=5, height=4, res=150, printFontSize=10,
               saveFontSize=18, style="bw")  #set a few options

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{bmp,}
\keyword{ggplot2,}
\keyword{jpeg,}
\keyword{pdf}
\keyword{png,}
\keyword{tiff,}
