% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printAndSave.R
\name{printAndSave}
\alias{printAndSave}
\title{Function  printAndSave}
\usage{
printAndSave(
  plotObject,
  filename,
  width = 7,
  height = 5,
  units = "in",
  res = 300,
  scale = 1,
  printFontSize = 12,
  saveFontSize = 24,
  scaleLegend = TRUE,
  printPlot = TRUE,
  savePlot = TRUE
)
}
\arguments{
\item{plotObject}{A ggplot2 plot object}

\item{filename}{A path/file name for the graphic}

\item{width}{Graphic width in inches (Default = 7)}

\item{height}{Graphic height in inches (Default = 5)}

\item{units}{Units for height and width ("in"|"cm"|"mm") (Default = "in")}

\item{res}{Resolution in PPI (Default = 300)}

\item{scale}{Multiplicative scaling factor (Default = 1)}

\item{printFontSize}{Base font size for the graphic on the console/knitr (Default = 12)}

\item{saveFontSize}{Base font size for the graphic file (Default = 24)}

\item{scaleLegend}{Scale the legend smaller if font > 14  (Default = TRUE)}

\item{printPlot}{Print to console if TRUE (Default = TRUE)}

\item{savePlot}{Print to file if TRUE (Default = TRUE)}
}
\value{
The print object
}
\description{
printAndSave is a wrapper around ggsave.  It prints a ggplot2 object to the
console or a knitr report and saves as it a graphic file. By default it uses larger fonts
for the graphic file (more suitable for PPT).  Graphic file type is taken
from the given filename and can be one of: png, bmp, tiff, jpg, pdf, or svg.
}
\details{
The function increases the size of the print/knitr legend and
decreases the size of the legend in the saved
graphic file.  This works well for images of about 5-7 inches but may not scale
well to other image sizes.  Set scaleLegend to FALSE to get the default behavior
back or use legend.key.size in a theme to tweak to your own specifications.
}
\examples{
\dontrun{
   # All defaults
   printAndSave(Myggplot, "myfile.png")
   # Set a few options
   printAndSave(Myggplot,
                "myfile.png",
                width = 5,
                height = 4,
                res = 150,
                printFontSize = 10,
                saveFontSize = 18)
}

}
