% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printAndSave.R
\name{printWithFootnote}
\alias{printWithFootnote}
\title{Function  printWithFootnote}
\usage{
printWithFootnote(
  plotObject,
  footnote,
  fontface = "plain",
  fontsize = 10,
  hjust = -0.1
)
}
\arguments{
\item{plotObject}{A ggplot2 plotobject}

\item{footnote}{A path/filename for the graphic}

\item{fontface}{fontface for the footnotw (default = "plain")}

\item{fontsize}{size of the footnote font (default = 10)}

\item{hjust}{Specify horizontal justification (Default = -0.1)}
}
\value{
Prints the graphic object to the console
}
\description{
Print a ggplot2 object to the console/knitr report adding footnote text
under the plot.  Use when you want the footnot to be underneath the plot
labels.  Only prints the footnote once on a facetted plot.
}
\examples{
    #Write to the console or knitr report
    printWithFootnote(Myggplot, footnote = "Footnote Text")

    #Capture to a file
    png ("myplot.png", width=5, height=4, unit="in")
    printWithFootnote(Myggplot, footnote = "Footnote Text")
    invisible(dev.off())

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{bmp,}
\keyword{ggplot2,}
\keyword{jpeg,}
\keyword{pdf}
\keyword{png,}
\keyword{tiff,}
