% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printAndSave.R
\name{printWithFootnote}
\alias{printWithFootnote}
\title{Function  printWithFootnote}
\usage{
printWithFootnote(
  plotObject,
  footnote,
  fontface = "plain",
  fontsize = 10,
  hjust = -0.1
)
}
\arguments{
\item{plotObject}{A ggplot2 plot object}

\item{footnote}{A path/file name for the graphic}

\item{fontface}{Font face for the footnote (Default = "plain")}

\item{fontsize}{Size of the footnote font (Default = 10)}

\item{hjust}{Specify horizontal justification (Default = -0.1)}
}
\value{
Prints the graphic object to the console
}
\description{
Print a ggplot2 object to the console or a knitr report adding footnote text
under the plot.  Use when you want the footnote to be underneath the plot
labels.  Only prints the footnote once on a faceted plot.
}
\examples{
\dontrun{
    # Write to the console or knitr report
    printWithFootnote(Myggplot, footnote = "Footnote Text")

    # Capture to a file
    png("myplot.png", width = 5, height = 4, unit = "in")
    printWithFootnote(Myggplot, footnote = "Footnote Text")
    invisible(dev.off())
}

}
