% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdgeRNorm.R
\name{runEdgeRNorm}
\alias{runEdgeRNorm}
\title{Function  runEdgeRNorm (edgeR normlization )}
\usage{
runEdgeRNorm(dat, NormMethod = "TMM", plotFile = "TMM_Norm.Factors.PNG",
  plotLabels = NULL)
}
\arguments{
\item{dat}{A DGEobj or RSE object data structure containing counts and design
data}

\item{plotFile}{Enable a Barplot of the norm.factors produced (Default = "Norm.Factors.PNG")
Set to NULL to disable the plot.}

\item{plotLabels}{Sample labels for the plot. Length must equal the number of
samples. (Default = NULL;  sample number will be displayed)}

\item{normMethod}{One of "TMM", "RLE", "upperquartile" or "none". Default = "TMM"}
}
\value{
A DGEObj with a normalized DGEList added.
}
\description{
Returns a SubsettableListOfArrays (SLOA) containing the Fit and associated
output from the standard RNA-Seq using edgeR TMM normalization
followed by limma voom and lmfit.  The SLOA contains the Fit object
and various associated data  use "print(SLOA)" to examine the contents.
}
\examples{
MyDgeObj <- runEdgeRNorm (RSE)

MyDgeObj <- runEdgeRNorm (MyDgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Entrez,}
\keyword{GeneID}
\keyword{gene}
\keyword{symbol,}

