% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdgeRNorm.R
\name{runEdgeRNorm}
\alias{runEdgeRNorm}
\title{Run edgeR TMM normalization on DGEobj}
\usage{
runEdgeRNorm(
  dgeObj,
  normMethod = "TMM",
  plotFile = "TMM_Norm.Factors.PNG",
  plotLabels = NULL
)
}
\arguments{
\item{dgeObj}{A DGEobj containing counts, design data, and gene annotation.}

\item{normMethod}{One of "TMM", "RLE", "upperquartile", or "none". (Default = "TMM")}

\item{plotFile}{Enable a bar plot of the norm.factors produced (Default = "Norm.Factors.PNG")
Set to NULL to disable the plot.}

\item{plotLabels}{Sample labels for the plot. Length must equal the number of
samples. (Default = NULL; sample number will be displayed)}
}
\value{
A DGEobj with a normalized DGEList added.
}
\description{
Returns a DGEobj containing DGEList object representing the result of
 edgeR TMM normalization.
}
\examples{
\dontrun{
   myDGEobj <- runEdgeRNorm(myDGEobj)
}

}
