% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSVA.R
\name{runSVA}
\alias{runSVA}
\title{Function  runSVA}
\usage{
runSVA(MySLOA)
}
\arguments{
\item{MySLOA}{An SLOA object with normalized counts and a DesignMatrix.}
}
\value{
MySLOA An SLOA object containing data ready for DGE analysis (runVoom)
}
\description{
Takes an SubsettableListOfArrays (SLOA) object from runVoom and returns
the SLOA object after running SVA and adding the surrogate variables it
discovers to the DesignMatrix in the SLOA.  The resulting SLOA object is
ready to be reprocessed by runVoom again to complete the analysis.
}
\examples{
MySLOA = runSVA (MySLOA)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{SLOA}
\keyword{SVA}

