% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSVA.R
\name{runSVA}
\alias{runSVA}
\title{Function  runSVA}
\usage{
runSVA(dgeObj, designMatrixName)
}
\arguments{
\item{dgeObj}{A DGEobj with normalized counts and a DesignMatrix.}
}
\value{
dgeObj The DGEobj is returned containing a new design matrix.
}
\description{
Takes an DGEobj from runVoom and tests for surrogate variables.  Adds a new
design matrix to the DGEobj with the surrogate variable columns appended (cbind). 
runVoom should then be run again with the new design matrix to complete the
analysis.
}
\examples{
MyDgeObj = runSVA (MyDgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{SLOA}
\keyword{SVA}

