% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSVA.R
\name{runSVA}
\alias{runSVA}
\title{Function  runSVA}
\usage{
runSVA(dgeObj, designMatrixName)
}
\arguments{
\item{MySLOA}{An SLOA object with normalized counts and a DesignMatrix.}
}
\value{
MySLOA An SLOA object containing data ready for DGE analysis (runVoom)
}
\description{
Takes an DGEobj from runVoom and tests for surrogate variables.  Adds a new
design matrix to the DGEobj with the surrogate varaible columns appended. 
runVoom should then be run again with the new design matrix to complete the
analysis.
}
\examples{
MySLOA = runSVA (MySLOA)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{SLOA}
\keyword{SVA}

