% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVoom.R
\name{runVoom}
\alias{runVoom}
\title{Function  runVoom (voom then lmFit)}
\usage{
runVoom(dgelist, designMatrix, formula, dupcorBlock = NULL,
  qualityWeights = TRUE, var.design = NULL, dgeResult = NULL)
}
\arguments{
\item{dgelist}{An edgeR DGEList object that has already been normalized 
(e.g. by edgeR::calcNormFactors or DGE.Tools::runEdgeRNorm) (required)}

\item{designMatrix}{A design matrix created by model.matrix (required)}

\item{formula}{A text representation of a formula to fit (required)}

\item{dupcorBlock}{Supply a block argument to trigger duplicateCorrelation (optional)}

\item{qualityWeights}{Runs VoomWithQualityWeights if set to TRUE (default=TRUE)}

\item{var.design}{Provide a design matrix to identify replicate groups.  If omitted
VoomWithQualityWeights treats each sample individually.}

\item{dgeResult}{A List object to add results to. If supplied, results 
will be added to this. Otherwise a new dgeResult will be created. This
allows you to sequentially capture results of multiple steps by passing
the dgeResult to each function in the pipeline that supports it.}
}
\value{
A dgeResult object (currently a simple List of Lists structure)
}
\description{
Input is minimally a DGEList (typically from counts %>% DGEList %>%
calcNormFactors), a design matrix (from model.matrix) and a formula (text
representation).  Instead of a dgelist, you can probably use a Log2CPM matrix
(preferably already normalized).
}
\details{
Returns a dgeResult object containing the VoomElist and Fit objects derived from
voom and lmfit respectively. Appends to a prior dgeResult if one is passed, else
creates a new dgeResult.
}
\examples{

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{Entrez,}
\keyword{GeneID}
\keyword{gene}
\keyword{symbol,}

