% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyContrasts.R
\name{tidyContrasts}
\alias{tidyContrasts}
\title{Function  tidyContrasts}
\usage{
tidyContrasts(x, rownameColumn = "rownames", includeColumns)
}
\arguments{
\item{x}{A DGEobj or named list of contrast dataframes
(required).}

\item{rownameColumn}{Name of the rowname column. If a column by this
name does not exist, it is created from the rownames property (rownames_to_column(var=rownameColumn))}

\item{includeColumns}{A character vector of columns to include in the output (default = colnames of the first contrast)}
}
\value{
A DF with merged contrast data.
}
\description{
Takes a DGEobj or contrast list as input and merges them into one tidy dataframe.
A contrast list is simply a list of topTable contrast dataframes. For
example, DGEobj::getType(mydgeobj, "topTable") would retrieve a list of
contrasts from a DGEobj.  The contrast list must be a named list as the
contrast names are used during the merge operation.
}
\details{
The input may or may not have rownames. If supplied rownameColumn does not exist as a colname in
the dataframes, it is created from the rownames.  In tidy style, the output will have no rownames.

The contrast names will be used as a new column in the tidy output format.
}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq;}
\keyword{contrasts;}
\keyword{merge}
\keyword{tidy}
