% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyIntensity.R
\name{tidyIntensity}
\alias{tidyIntensity}
\title{Function  tidyIntensity}
\usage{
tidyIntensity(x, rowidColname, keyColname, valueColname, group)
}
\arguments{
\item{x}{A dataframe or matrix with rownames and colnames (required).}

\item{rowidColname}{Column name for the rowID (Usually a geneID column).
If it doesn't exist, it will be created from rownames.}

\item{valueColname}{Column name for the value column}

\item{group}{A grouping variable to define which columns go together (vector
with length = ncol(x); coercable to character)}
}
\value{
A tidy dataframe of intensity data.
}
\description{
Takes dataframe or matrix of intensity values with geneid rownames and samplename columnnames and a grouping vector.
Returns a tidy dataframe.
}
\examples{

  MyTidyIntensity  = tidyIntensity (MyLog2CPM)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq;}
\keyword{intensity;}
\keyword{tidy}
