% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCounts.R
\name{tpm}
\alias{tpm}
\title{Function tpm}
\usage{
tpm(dgeo, applyFilter = TRUE)
}
\arguments{
\item{dgeo}{A DGEobj data structure}

\item{applyFilter}{If TRUE, reduces to the filtered gene list. FALSE returns
all genes in the raw data. [Default = TRUE]}
}
\value{
A matrix in the new unit space
}
\description{
Takes a DGEObj as input. Uses all data (pre-genefiltering; counts_orig) to
calculate TPM values. Then optionally filters any genes that were already
filtered out.
  
For GeneLength in the calculation, it takes geneData$ExonLength (for Omicsoft data) 
or uses rowMeans(effectiveLength) for data derived from Xpress.
}
\details{
Internally it uses edgeR::fpkm to calculate fpkm and converts to tpm
using the formula provided  by [Harold Pimental](https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/).
}
\examples{
myTPM <- tpm(dgeobj) 

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq,}
\keyword{conversion}
\keyword{unit}
