% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voomWorkflow.R
\name{voomWorkflow}
\alias{voomWorkflow}
\title{Function voomWorkflow}
\usage{
voomWorkflow(dgeObj, formula, projectName, designMatrixName,
  fracThreshold = 0.5, outputPath = "./", annotationFile)
}
\arguments{
\item{dgeObj}{A class dgeObj with counts, gene annotation and sample annotation}

\item{formula}{A text representation of the formula you want to use}

\item{projectName}{This should be the project name from Xpress or Omicsoft}

\item{designMatrixName}{User defined name for the design matrix}

\item{fracThreshold}{Fraction of samples that must meet intensity thresholds to keep a gene (Default = 0.5)}

\item{outputPath}{Where to send output plots}

\item{annotationFile}{File of key=value pairs to populate required attributes (optional but highly advised)}
}
\value{
A DGEobj with a modified object definition embedded.
}
\description{
This function runs several steps in the RNA-Seq pipeline in one fell swoop.
It supports duplicateCorrelation if you provide a blocking vector.  It does
not directly support SVA.
}
\details{
For SVA, you need to run SVA first and add the SVA variables to your design
table. Then you can proceed with this function.
}
\examples{

   MyDgeObj <- newType(MyDgeObj,

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
